Formula wrapper for crr from cmprsk

Formula interface for Fine-Gray regression competing risk models.

Usage

FGR(formula, data, cause = 1, y = TRUE, ...)

Arguments

formula

A formula whose left hand side is a Hist object -- see Hist. The right hand side specifies (a linear combination of) the covariates. See examples below.

data

A data.frame in which all the variables of formula can be interpreted.

cause

The failure type of interest. Defaults to 1.

y

logical value: if TRUE, the response vector is returned in component response.

...

...

Value

See crr.

Details

Formula interface for the function crr from the cmprsk package.

The function crr allows to multiply some covariates by time before they enter the linear predictor. This can be achieved with the formula interface, however, the code becomes a little cumbersome. See the examples. Note that FGR does not allow for delayed entry (left-truncation).

References

Gerds, TA and Scheike, T and Andersen, PK (2011) Absolute risk regression for competing risks: interpretation, link functions and prediction Research report 11/7. Department of Biostatistics, University of Copenhagen

See also

Author

Thomas Alexander Gerds tag@biostat.ku.dk

Examples

library(prodlim)
library(survival)
library(cmprsk)
library(lava)
#> 
#> Attaching package: ‘lava’
#> The following objects are masked from ‘package:zoo’:
#> 
#>     index, index<-
d <- prodlim::SimCompRisk(100)
f1 <- FGR(Hist(time,cause)~X1+X2,data=d)
#> Argument cause missing. Analyse cause: 1
print(f1)
#> 
#> Right-censored response of a competing.risks model
#> 
#> No.Observations: 100 
#> 
#> Pattern:
#>          
#> Cause     event right.censored
#>   1          50              0
#>   2          29              0
#>   unknown     0             21
#> 
#> 
#> Fine-Gray model: analysis of cause 1 
#> 
#> Competing Risks Regression
#> 
#> Call:
#> FGR(formula = Hist(time, cause) ~ X1 + X2, data = d, cause = "1")
#> 
#>     coef exp(coef) se(coef)    z p-value
#> X1 0.834      2.30    0.309 2.70 6.9e-03
#> X2 0.723      2.06    0.137 5.28 1.3e-07
#> 
#>    exp(coef) exp(-coef) 2.5% 97.5%
#> X1      2.30      0.434 1.26  4.22
#> X2      2.06      0.485 1.58  2.69
#> 
#> Num. cases = 100
#> Pseudo Log-likelihood = -190 
#> Pseudo likelihood ratio test = 28.3  on 2 df,
#> 
#> Convergence: TRUE 
#> 

## crr allows that some covariates are multiplied by
## a function of time (see argument tf of crr)
## by FGR uses the identity matrix
f2 <- FGR(Hist(time,cause)~cov2(X1)+X2,data=d)
#> Argument cause missing. Analyse cause: 1
print(f2)
#> 
#> Right-censored response of a competing.risks model
#> 
#> No.Observations: 100 
#> 
#> Pattern:
#>          
#> Cause     event right.censored
#>   1          50              0
#>   2          29              0
#>   unknown     0             21
#> 
#> 
#> Fine-Gray model: analysis of cause 1 
#> 
#> Competing Risks Regression
#> 
#> Call:
#> FGR(formula = Hist(time, cause) ~ cov2(X1) + X2, data = d, cause = "1")
#> 
#>         coef exp(coef) se(coef)    z p-value
#> X2     0.740      2.10   0.1368 5.41 6.4e-08
#> X1*tf1 0.135      1.14   0.0613 2.21 2.7e-02
#> 
#>        exp(coef) exp(-coef) 2.5% 97.5%
#> X2          2.10      0.477 1.60  2.74
#> X1*tf1      1.14      0.873 1.02  1.29
#> 
#> Num. cases = 100
#> Pseudo Log-likelihood = -191 
#> Pseudo likelihood ratio test = 25.8  on 2 df,
#> 
#> Convergence: TRUE 
#> 

## same thing, but more explicit:
f3 <- FGR(Hist(time,cause)~cov2(X1)+cov1(X2),data=d)
#> Argument cause missing. Analyse cause: 1
print(f3)
#> 
#> Right-censored response of a competing.risks model
#> 
#> No.Observations: 100 
#> 
#> Pattern:
#>          
#> Cause     event right.censored
#>   1          50              0
#>   2          29              0
#>   unknown     0             21
#> 
#> 
#> Fine-Gray model: analysis of cause 1 
#> 
#> Competing Risks Regression
#> 
#> Call:
#> FGR(formula = Hist(time, cause) ~ cov2(X1) + cov1(X2), data = d, 
#>     cause = "1")
#> 
#>         coef exp(coef) se(coef)    z p-value
#> X2     0.740      2.10   0.1368 5.41 6.4e-08
#> X1*tf1 0.135      1.14   0.0613 2.21 2.7e-02
#> 
#>        exp(coef) exp(-coef) 2.5% 97.5%
#> X2          2.10      0.477 1.60  2.74
#> X1*tf1      1.14      0.873 1.02  1.29
#> 
#> Num. cases = 100
#> Pseudo Log-likelihood = -191 
#> Pseudo likelihood ratio test = 25.8  on 2 df,
#> 
#> Convergence: TRUE 
#> 

## both variables can enter cov2:
f4 <- FGR(Hist(time,cause)~cov2(X1)+cov2(X2),data=d)
#> Argument cause missing. Analyse cause: 1
print(f4)
#> 
#> Right-censored response of a competing.risks model
#> 
#> No.Observations: 100 
#> 
#> Pattern:
#>          
#> Cause     event right.censored
#>   1          50              0
#>   2          29              0
#>   unknown     0             21
#> 
#> 
#> Fine-Gray model: analysis of cause 1 
#> 
#> Competing Risks Regression
#> 
#> Call:
#> FGR(formula = Hist(time, cause) ~ cov2(X1) + cov2(X2), data = d, 
#>     cause = "1")
#> 
#>          coef exp(coef) se(coef)    z p-value
#> X1*tf1 0.1361      1.15   0.0611 2.23   0.026
#> X2*tf2 0.0985      1.10   0.0434 2.27   0.023
#> 
#>        exp(coef) exp(-coef) 2.5% 97.5%
#> X1*tf1      1.15      0.873 1.02  1.29
#> X2*tf2      1.10      0.906 1.01  1.20
#> 
#> Num. cases = 100
#> Pseudo Log-likelihood = -198 
#> Pseudo likelihood ratio test = 12.1  on 2 df,
#> 
#> Convergence: TRUE 
#> 

## change the function of time
qFun <- function(x){x^2}
noFun <- function(x){x}
sqFun <- function(x){x^0.5}

## multiply X1 by time^2 and X2 by time:
f5 <- FGR(Hist(time,cause)~cov2(X1,tf=qFun)+cov2(X2),data=d)
#> Argument cause missing. Analyse cause: 1
#> Error in qFun(c(0.223312599010185, 0.844653052602936, 0.86743957909998, 0.869654848917858, 1.4718454380646, 1.47986846377743, 1.56788544286003, 1.57284176822651, 1.71280654079967, 1.7446175554818, 2.08017045127398, 2.11900022748555, 2.34027091945108, 2.44735526652256, 2.47481572672078, 2.69072148768644, 2.69930631257876, 2.82681955875637, 2.86986677073091, 3.28835303045329, 3.38016548882937, 3.42930837661085, 3.57927210122299, 3.61310741873331, 3.71589696889436, 3.82638651755888, 4.13125614409393, 4.22817648821086, 4.69740197012358, 4.75934581546958, 4.7989296471193, 5.09144457759695, 5.20599159981632, 5.27409575450788, 5.37841476839829, 5.45365210472182, 5.47792252614271, 5.8344164585106, 6.03541407396753, 6.63251816105419, 6.92722896832721, 7.2931295159595, 7.59440495376659, 7.82755206098875, 8.10614220085668, 8.1127149986577, 8.83868607919211, 9.05983471886304, 9.33426459771812, 10.6845775000325)): could not find function "qFun"
print(f5)
#> Error in print(f5): object 'f5' not found
print(f5$crrFit)
#> Error in print(f5$crrFit): object 'f5' not found
## same results as crr
with(d,crr(ftime=time,
           fstatus=cause,
           cov2=d[,c("X1","X2")],
           tf=function(time){cbind(qFun(time),time)}))
#> convergence:  TRUE 
#> coefficients:
#>  X1*tf1 X2*time 
#> 0.01259 0.09674 
#> standard errors:
#> [1] 0.00853 0.04345
#> two-sided p-values:
#>  X1*tf1 X2*time 
#>   0.140   0.026 

## still same result, but more explicit
f5a <- FGR(Hist(time,cause)~cov2(X1,tf=qFun)+cov2(X2,tf=noFun),data=d)
#> Argument cause missing. Analyse cause: 1
#> Error in qFun(c(0.223312599010185, 0.844653052602936, 0.86743957909998, 0.869654848917858, 1.4718454380646, 1.47986846377743, 1.56788544286003, 1.57284176822651, 1.71280654079967, 1.7446175554818, 2.08017045127398, 2.11900022748555, 2.34027091945108, 2.44735526652256, 2.47481572672078, 2.69072148768644, 2.69930631257876, 2.82681955875637, 2.86986677073091, 3.28835303045329, 3.38016548882937, 3.42930837661085, 3.57927210122299, 3.61310741873331, 3.71589696889436, 3.82638651755888, 4.13125614409393, 4.22817648821086, 4.69740197012358, 4.75934581546958, 4.7989296471193, 5.09144457759695, 5.20599159981632, 5.27409575450788, 5.37841476839829, 5.45365210472182, 5.47792252614271, 5.8344164585106, 6.03541407396753, 6.63251816105419, 6.92722896832721, 7.2931295159595, 7.59440495376659, 7.82755206098875, 8.10614220085668, 8.1127149986577, 8.83868607919211, 9.05983471886304, 9.33426459771812, 10.6845775000325)): could not find function "qFun"
f5a$crrFit
#> Error in eval(expr, envir, enclos): object 'f5a' not found

## multiply X1 by time^2 and X2 by sqrt(time)
f5b <- FGR(Hist(time,cause)~cov2(X1,tf=qFun)+cov2(X2,tf=sqFun),data=d,cause=1)
#> Error in qFun(c(0.223312599010185, 0.844653052602936, 0.86743957909998, 0.869654848917858, 1.4718454380646, 1.47986846377743, 1.56788544286003, 1.57284176822651, 1.71280654079967, 1.7446175554818, 2.08017045127398, 2.11900022748555, 2.34027091945108, 2.44735526652256, 2.47481572672078, 2.69072148768644, 2.69930631257876, 2.82681955875637, 2.86986677073091, 3.28835303045329, 3.38016548882937, 3.42930837661085, 3.57927210122299, 3.61310741873331, 3.71589696889436, 3.82638651755888, 4.13125614409393, 4.22817648821086, 4.69740197012358, 4.75934581546958, 4.7989296471193, 5.09144457759695, 5.20599159981632, 5.27409575450788, 5.37841476839829, 5.45365210472182, 5.47792252614271, 5.8344164585106, 6.03541407396753, 6.63251816105419, 6.92722896832721, 7.2931295159595, 7.59440495376659, 7.82755206098875, 8.10614220085668, 8.1127149986577, 8.83868607919211, 9.05983471886304, 9.33426459771812, 10.6845775000325)): could not find function "qFun"

## additional arguments for crr
f6<- FGR(Hist(time,cause)~X1+X2,data=d, cause=1,gtol=1e-5)
f6
#> 
#> Right-censored response of a competing.risks model
#> 
#> No.Observations: 100 
#> 
#> Pattern:
#>          
#> Cause     event right.censored
#>   1          50              0
#>   2          29              0
#>   unknown     0             21
#> 
#> 
#> Fine-Gray model: analysis of cause 1 
#> 
#> Competing Risks Regression
#> 
#> Call:
#> FGR(formula = Hist(time, cause) ~ X1 + X2, data = d, cause = 1, 
#>     gtol = 1e-05)
#> 
#>     coef exp(coef) se(coef)    z p-value
#> X1 0.834      2.30    0.309 2.70 6.9e-03
#> X2 0.723      2.06    0.137 5.28 1.3e-07
#> 
#>    exp(coef) exp(-coef) 2.5% 97.5%
#> X1      2.30      0.434 1.26  4.22
#> X2      2.06      0.485 1.58  2.69
#> 
#> Num. cases = 100
#> Pseudo Log-likelihood = -190 
#> Pseudo likelihood ratio test = 28.3  on 2 df,
#> 
#> Convergence: TRUE 
#> 
f6a<- FGR(Hist(time,cause)~X1+X2,data=d, cause=1,gtol=0.1)
f6a
#> 
#> Right-censored response of a competing.risks model
#> 
#> No.Observations: 100 
#> 
#> Pattern:
#>          
#> Cause     event right.censored
#>   1          50              0
#>   2          29              0
#>   unknown     0             21
#> 
#> 
#> Fine-Gray model: analysis of cause 1 
#> 
#> Competing Risks Regression
#> 
#> Call:
#> FGR(formula = Hist(time, cause) ~ X1 + X2, data = d, cause = 1, 
#>     gtol = 0.1)
#> 
#>     coef exp(coef) se(coef)    z p-value
#> X1 0.895      2.45    0.316 2.83 4.7e-03
#> X2 0.774      2.17    0.141 5.48 4.1e-08
#> 
#>    exp(coef) exp(-coef) 2.5% 97.5%
#> X1      2.45      0.409 1.32  4.55
#> X2      2.17      0.461 1.64  2.86
#> 
#> Num. cases = 100
#> Pseudo Log-likelihood = -190 
#> Pseudo likelihood ratio test = 28.1  on 2 df,
#> 
#> Convergence: TRUE 
#>