Explained variation for settings with binary, survival and competing risk outcome

Index of Prediction Accuracy: General R^2 for binary outcome and right censored time to event (survival) outcome also with competing risks

Usage

rsquared(object,...)
IPA(object,...)
# S3 method for default
rsquared(object,formula,newdata,times,cause,...)
# S3 method for glm
rsquared(object,formula,newdata,...)
# S3 method for coxph
rsquared(object,formula,newdata,times,...)
# S3 method for CauseSpecificCox
rsquared(object,formula,newdata,times,cause,...)
# S3 method for default
IPA(object,formula,newdata,times,cause,...)
# S3 method for glm
IPA(object,formula,newdata,...)
# S3 method for coxph
IPA(object,formula,newdata,times,...)
# S3 method for CauseSpecificCox
IPA(object,formula,newdata,times,cause,...)

Arguments

object

Model for which we want IPA.

...

passed to riskRegression::Score

newdata

Optional validation data set in which to compute IPA

formula

Formula passed to Score. If not provided, try to use the formula of the call of object, if any.

cause

For competing risk models the event of interest

times

Vector of time points used as prediction horizon for the computation of Brier scores.

Value

Data frame with explained variation values for the full model.

Details

IPA (R^2) is calculated based on the model's predicted risks. The Brier score of the model is compared to the Brier score of the null model.

See also

Author

Thomas A. Gerds <tag@biostat.ku.dk>

Examples

library(prodlim)
library(data.table)
# binary outcome
library(lava)
set.seed(18)
learndat <- sampleData(48,outcome="binary")
lr1 = glm(Y~X1+X2+X7+X9,data=learndat,family=binomial)
IPA(lr1)
#> Error in eval(object$call$data): object 'learndat' not found

## validation data
valdat=sampleData(94,outcome="binary")
IPA(lr1,newdata=valdat)
#>      Variable Brier  IPA IPA.drop
#> 1: Null model  23.6  0.0     -1.3
#> 2: Full model  23.9 -1.3      0.0
#> 3:         X1  23.9 -0.9     -0.3
#> 4:         X2  23.2  1.9     -3.1
#> 5:         X7  24.0 -1.7      0.4
#> 6:         X9  23.5  0.5     -1.8
#> 
#> NOTE: Values are multiplied by 100 and given in %.
#> NOTE: The higher IPA the better.
#> NOTE: IPA.drop = IPA(Full model) - IPA. The higher the drop
#> the more important is the variable for the full model.

## predicted risks externally given
p1=predictRisk(lr1,newdata=valdat)
IPA(p1,formula=Y~1,valdat)
#>         model Brier  IPA
#> 1: Null model  23.6  0.0
#> 2:    numeric  23.9 -1.3
#> 
#> NOTE: Values are multiplied by 100 and given in %.
#> NOTE: The higher IPA the better.
#> NOTE: IPA.drop = IPA(Full model) - IPA. The higher the drop
#> the more important is the variable for the full model.

# survival
library(survival)
data(pbc)
pbc=na.omit(pbc)
pbctest=(1:NROW(pbc)) %in% sample(1:NROW(pbc),size=.632*NROW(pbc))
pbclearn=pbc[pbctest,]
cox1= coxph(Surv(time,status!=0)~age+sex+log(bili)+log(albumin)+log(protime),
      data=pbclearn,x=TRUE)

## same data
IPA(cox1,formula=Surv(time,status!=0)~1,times=1000)
#> Error in eval(object$call$data): object 'pbclearn' not found

## validation data
pbcval=pbc[!pbctest,]
IPA(cox1,formula=Surv(time,status!=0)~1,newdata=pbcval,times=1000)
#> Warning: Invalid .internal.selfref detected and fixed by taking a (shallow) copy of the data.table so that := can add this new column by reference. At an earlier point, this data.table has been copied by R (or was created manually using structure() or similar). Avoid names<- and attr<- which in R currently (and oddly) may copy the whole data.table. Use set* syntax instead to avoid copying: ?set, ?setnames and ?setattr. If this message doesn't help, please report your use case to the data.table issue tracker so the root cause can be fixed or this message improved.
#>        Variable times Brier  IPA IPA.drop
#> 1:   Null model  1000  18.5  0.0     25.1
#> 2:   Full model  1000  13.9 25.1      0.0
#> 3:          age  1000  14.0 24.4      0.6
#> 4:          sex  1000  13.9 24.8      0.3
#> 5:    log(bili)  1000  15.2 18.0      7.1
#> 6: log(albumin)  1000  13.7 25.9     -0.8
#> 7: log(protime)  1000  14.2 23.4      1.7
#> 
#> NOTE: Values are multiplied by 100 and given in %.
#> NOTE: The higher IPA the better.
#> NOTE: IPA.drop = IPA(Full model) - IPA. The higher the drop
#> the more important is the variable for the full model.

## predicted risks externally given
p2=predictRisk(cox1,newdata=pbcval,times=1000)
IPA(cox1,formula=Surv(time,status!=0)~1,newdata=pbcval,times=1000)
#> Warning: Invalid .internal.selfref detected and fixed by taking a (shallow) copy of the data.table so that := can add this new column by reference. At an earlier point, this data.table has been copied by R (or was created manually using structure() or similar). Avoid names<- and attr<- which in R currently (and oddly) may copy the whole data.table. Use set* syntax instead to avoid copying: ?set, ?setnames and ?setattr. If this message doesn't help, please report your use case to the data.table issue tracker so the root cause can be fixed or this message improved.
#>        Variable times Brier  IPA IPA.drop
#> 1:   Null model  1000  18.5  0.0     25.1
#> 2:   Full model  1000  13.9 25.1      0.0
#> 3:          age  1000  14.0 24.4      0.6
#> 4:          sex  1000  13.9 24.8      0.3
#> 5:    log(bili)  1000  15.2 18.0      7.1
#> 6: log(albumin)  1000  13.7 25.9     -0.8
#> 7: log(protime)  1000  14.2 23.4      1.7
#> 
#> NOTE: Values are multiplied by 100 and given in %.
#> NOTE: The higher IPA the better.
#> NOTE: IPA.drop = IPA(Full model) - IPA. The higher the drop
#> the more important is the variable for the full model.
 
# competing risks
data(Melanoma)
Melanomatest=(1:NROW(Melanoma)) %in% sample(1:NROW(Melanoma),size=.632*NROW(Melanoma))
Melanomalearn=Melanoma[Melanomatest,]
fit1 <- CSC(list(Hist(time,status)~sex,
                 Hist(time,status)~invasion+epicel+age),
                 data=Melanoma)
IPA(fit1,times=1000,cause=2)
#>      Variable times Brier  IPA IPA.drop
#> 1: Null model  1000   3.3  0.0      2.9
#> 2: Full model  1000   3.2  2.9      0.0
#> 3:        sex  1000   3.2  2.8      0.1
#> 4:   invasion  1000   3.2  4.0     -1.2
#> 5:     epicel  1000   3.2  3.1     -0.2
#> 6:        age  1000   3.3 -0.8      3.6
#> 
#> NOTE: Values are multiplied by 100 and given in %.
#> NOTE: The higher IPA the better.
#> NOTE: IPA.drop = IPA(Full model) - IPA. The higher the drop
#> the more important is the variable for the full model.

## validation data
Melanomaval=Melanoma[!Melanomatest,]
IPA(fit1,formula=Hist(time,status)~1,newdata=Melanomaval,times=1000)
#>      Variable times Brier  IPA IPA.drop
#> 1: Null model  1000  11.6  0.0      4.6
#> 2: Full model  1000  11.0  4.6      0.0
#> 3:        sex  1000  11.6 -0.0      4.6
#> 4:   invasion  1000  11.0  4.4      0.2
#> 5:     epicel  1000  11.0  4.5      0.1
#> 6:        age  1000  11.0  4.6     -0.0
#> 
#> NOTE: Values are multiplied by 100 and given in %.
#> NOTE: The higher IPA the better.
#> NOTE: IPA.drop = IPA(Full model) - IPA. The higher the drop
#> the more important is the variable for the full model.

## predicted risks externally given
p3= predictRisk(fit1,cause=1,newdata=Melanomaval,times=1000)
IPA(p3,formula=Hist(time,status)~1,cause=1,newdata=Melanomaval,times=1000)
#>         model times Brier IPA
#> 1: Null model  1000  11.6 0.0
#> 2:     matrix  1000  11.0 4.6
#> 
#> NOTE: Values are multiplied by 100 and given in %.
#> NOTE: The higher IPA the better.
#> NOTE: IPA.drop = IPA(Full model) - IPA. The higher the drop
#> the more important is the variable for the full model.