Average Treatment Effects Computation

Use the g-formula or the IPW or the double robust estimator to estimate the average treatment effect (absolute risk difference or ratio) based on Cox regression with or without competing risks.

Usage

ate(
  event,
  treatment,
  censor = NULL,
  data,
  data.index = NULL,
  formula,
  estimator = NULL,
  strata = NULL,
  contrasts = NULL,
  allContrasts = NULL,
  times,
  cause = NA,
  landmark,
  se = TRUE,
  iid = (B == 0) && (se || band),
  known.nuisance = FALSE,
  band = FALSE,
  B = 0,
  seed,
  handler = "foreach",
  mc.cores = 1,
  cl = NULL,
  verbose = TRUE,
  ...
)

Arguments

event

Outcome model which describes how the probability of experiencing a terminal event depends on treatment and covariates. The object carry its own call and have a predictRisk method. See examples.

treatment

Treatment model which describes how the probability of being allocated to a treatment group depends on covariates. The object must be a glm object (logistic regression) or the name of the treatment variable. See examples.

censor

Censoring model which describes how the probability of being censored depends on treatment and covariates. The object must be a coxph or cph object. See examples.

data

[data.frame or data.table] Data set in which to evaluate risk predictions based on the outcome model

data.index

[numeric vector] Position of the observation in argument data relative to the dataset used to obtain the argument event, treatment, censor. Only necessary for the standard errors when computing the Average Treatment Effects on a subset of the data set.

formula

For analyses with time-dependent covariates, the response formula. See examples.

estimator

[character] The type of estimator used to compute the average treatment effect. Can be "G-formula", "IPTW", or "AIPTW". When using estimator="G-formula", a model for the outcome should be provided (argument event). When using estimator="IPTW", a model for the treatment should be provided (argument treatment), as well as for the censoring (if any, argument censor). When using estimator="AIPTW" (double robust estimator), a model for the outcome and the treatment should be provided (argument event and treatment), as well as for the censoring (if any, argument censor).

strata

[character] Strata variable on which to compute the average risk. Incompatible with treatment. Experimental.

contrasts

[character vector] levels of the treatment variable for which the risks should be assessed and compared. Default is to consider all levels.

allContrasts

[2-row character matrix] levels of the treatment variable to be compared. Default is to consider all pairwise comparisons.

times

[numeric vector] Time points at which to evaluate average treatment effects.

cause

[integer/character] the cause of interest.

landmark

for models with time-dependent covariates the landmark time(s) of evaluation. In this case, argument time may only be one value and for the prediction of risks it is assumed that that the covariates do not change between landmark and landmark+time.

se

[logical] If TRUE compute and add the standard errors to the output.

iid

[logical] If TRUE compute and add the influence function to the output.

known.nuisance

[logical] If FALSE the uncertainty related to the estimation of the nuisance parameters is ignored. This greatly simplifies computations but requires to use a double robust estimator. The resulting standard error is known to be consistent when all event, treatment, and censoring models are valid.

band

[logical] If TRUE compute and add the quantiles for the confidence bands to the output.

B

[integer, >0] the number of bootstrap replications used to compute the confidence intervals. If it equals 0, then the influence function is used to compute Wald-type confidence intervals/bands.

seed

[integer, >0] sed number used to generate seeds for bootstrap and to achieve reproducible results.

handler

[character] Parallel handler for bootstrap. "foreach" is the default and the only option on Windows. It uses parallel to create a cluster. Other operating systems can use "mclapply". This argument is ignored when mc.cores=1 and cl=NULL.

mc.cores

[integer, >0] The number of cores to use, i.e., the upper limit for the number of child processes that run simultaneously. Passed to parallel::mclapply or parallel::makeCluster. The option is initialized from environment variable mc_cores if set.

cl

A parallel socket cluster used to perform cluster calculation in parallel (output by parallel::makeCluster). The packages necessary to run the computations (e.g. riskRegression) must already be loaded on each worker. Only used when handler="foreach".

verbose

[logical] If TRUE inform about estimated run time.

...

passed to predictRisk

References

Brice Maxime Hugues Ozenne, Thomas Harder Scheike, Laila Staerk, and Thomas Alexander Gerds. On the estimation of average treatment effects with right- censored time to event outcome and competing risks. Biometrical Journal, 62 (3):751--763, 2020.

See also

Author

Brice Ozenne broz@sund.ku.dk and Thomas Alexander Gerds tag@biostat.ku.dk

Examples

library(survival)
library(rms)
library(prodlim)
library(data.table)
set.seed(10)

#### Survival settings  ####
#### ATE with Cox model ####

## generate data
n <- 100
dtS <- sampleData(n, outcome="survival")
dtS$time <- round(dtS$time,1)
dtS$X1 <- factor(rbinom(n, prob = c(0.3,0.4) , size = 2), labels = paste0("T",0:2))

## estimate the Cox model
fit <- cph(formula = Surv(time,event)~ X1+X2,data=dtS,y=TRUE,x=TRUE)

## compute the ATE at times 5, 6, 7, and 8 using X1 as the treatment variable
## standard error computed using the influence function
## confidence intervals / p-values based on asymptotic results
ateFit1a <- ate(fit, data = dtS, treatment = "X1", times = 5:8)
#>  Input variables 
#>  - Treatment            : X1 (3 levels: "T0" "T1" "T2")
#>  - Event                : event (cause: 1, censoring: 0)
#>  - Time  [min;max]      : time [0.6;15.7]
#>  - Eval. time           :  5  6  7  8
#>       number at risk T0   16 13 10  7
#>       number at risk T1   26 21 18 11
#>       number at risk T2    7  3  3  1
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#>  - Uncertainty: Gaussian approximation 
#>                 where the variance is estimated via the influence function 
#> 
#>  Processing
#>  - Prepare influence function: outcome done 
#>  - Point estimation: done 
#>  - Decomposition iid: done
#>  - Confidence intervals: done
summary(ateFit1a)
#>      Average treatment effect 
#> 
#>  - Treatment            : X1 (3 levels: "T0" "T1" "T2")
#>  - Event                : event (cause: 1, censoring: 0)
#>  - Time  [min;max]      : time [0.6;15.7]
#>  - Eval. time           :  5  6  7  8
#>       number at risk T0   16 13 10  7
#>       number at risk T1   26 21 18 11
#>       number at risk T2    7  3  3  1
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#>  - Uncertainty: Gaussian approximation 
#>                 where the variance is estimated via the influence function 
#> 
#>  Testing procedure
#>  - Null hypothesis     : given two treatments (A,B) and a specific timepoint, equal risks 
#>  - Confidence level    : 0.95
#> 
#>  Results: 
#>  - Difference in standardized risk (B-A) between time zero and 'time' 
#>                 reported on the scale [-1;1] (difference between two probabilities)
#>  (difference in average risks when treating all subjects with the experimental treatment (B),
#>                                 vs. treating all subjects with the reference treatment (A))
#> 
#>  time X1=A risk(X1=A) X1=B risk(X1=B) difference           ci p.value
#>     5   T0      0.472   T1      0.389    -0.0825 [-0.25;0.09]   0.341
#>     6   T0      0.550   T1      0.461    -0.0896 [-0.27;0.09]   0.341
#>     7   T0      0.609   T1      0.516    -0.0930 [-0.28;0.10]   0.339
#>     8   T0      0.641   T1      0.547    -0.0941 [-0.29;0.10]   0.338
#>     5   T0      0.472   T2      0.448    -0.0239 [-0.26;0.21]   0.841
#>     6   T0      0.550   T2      0.525    -0.0256 [-0.28;0.22]   0.841
#>     7   T0      0.609   T2      0.583    -0.0263 [-0.28;0.23]   0.841
#>     8   T0      0.641   T2      0.615    -0.0264 [-0.29;0.23]   0.841
#>     5   T1      0.389   T2      0.448     0.0586 [-0.17;0.29]   0.612
#>     6   T1      0.461   T2      0.525     0.0639 [-0.18;0.31]   0.611
#>     7   T1      0.516   T2      0.583     0.0667 [-0.19;0.32]   0.609
#>     8   T1      0.547   T2      0.615     0.0677 [-0.19;0.33]   0.608
#> 
#>  difference      : estimated difference in standardized risks 
#>  ci              : pointwise confidence intervals 
#>  p.value         : (unadjusted) p-value 
summary(ateFit1a, short = TRUE, type = "meanRisk")
#>  - Standardized risk between time zero and 'time'
#> 
#>  time X1  risk          ci
#>     5 T0 0.472 [0.33;0.62]
#>     6 T0 0.550 [0.40;0.70]
#>     7 T0 0.609 [0.46;0.76]
#>     8 T0 0.641 [0.49;0.79]
#>     5 T1 0.389 [0.26;0.52]
#>     6 T1 0.461 [0.33;0.59]
#>     7 T1 0.516 [0.38;0.66]
#>     8 T1 0.547 [0.41;0.69]
#>     5 T2 0.448 [0.25;0.65]
#>     6 T2 0.525 [0.31;0.74]
#>     7 T2 0.583 [0.36;0.80]
#>     8 T2 0.615 [0.39;0.84]
#> 
summary(ateFit1a, short = TRUE, type = "diffRisk")
#>  - Difference in standardized risk (B-A) between time zero and 'time' 
#> 
#>  time X1=A X1=B difference           ci p.value
#>     5   T0   T1    -0.0825 [-0.25;0.09]   0.341
#>     6   T0   T1    -0.0896 [-0.27;0.09]   0.341
#>     7   T0   T1    -0.0930 [-0.28;0.10]   0.339
#>     8   T0   T1    -0.0941 [-0.29;0.10]   0.338
#>     5   T0   T2    -0.0239 [-0.26;0.21]   0.841
#>     6   T0   T2    -0.0256 [-0.28;0.22]   0.841
#>     7   T0   T2    -0.0263 [-0.28;0.23]   0.841
#>     8   T0   T2    -0.0264 [-0.29;0.23]   0.841
#>     5   T1   T2     0.0586 [-0.17;0.29]   0.612
#>     6   T1   T2     0.0639 [-0.18;0.31]   0.611
#>     7   T1   T2     0.0667 [-0.19;0.32]   0.609
#>     8   T1   T2     0.0677 [-0.19;0.33]   0.608
#> 
summary(ateFit1a, short = TRUE, type = "ratioRisk")
#>  - Ratio of standardized risks (B/A) between time zero and 'time' 
#> 
#>  time X1=A X1=B ratio          ci p.value
#>     5   T0   T1 0.825 [0.50;1.15]   0.293
#>     6   T0   T1 0.837 [0.53;1.14]   0.295
#>     7   T0   T1 0.847 [0.56;1.14]   0.299
#>     8   T0   T1 0.853 [0.58;1.13]   0.301
#>     5   T0   T2 0.949 [0.46;1.43]   0.838
#>     6   T0   T2 0.953 [0.51;1.40]   0.839
#>     7   T0   T2 0.957 [0.54;1.37]   0.839
#>     8   T0   T2 0.959 [0.56;1.36]   0.839
#>     5   T1   T2 1.150 [0.54;1.76]   0.628
#>     6   T1   T2 1.139 [0.58;1.69]   0.624
#>     7   T1   T2 1.129 [0.62;1.64]   0.622
#>     8   T1   T2 1.124 [0.63;1.61]   0.620
#> 

if (FALSE) {
## same as before with in addition the confidence bands / adjusted p-values
## (argument band = TRUE)
ateFit1b <- ate(fit, data = dtS, treatment = "X1", times = 5:8,
                band = TRUE)
summary(ateFit1b)

## by default bands/adjuste p-values computed separately for each treatment modality
summary(ateFit1b, band = 1,
         se = FALSE, type = "diffRisk", short = TRUE, quantile = TRUE)
## adjustment over treatment and time using the band argument of confint
summary(ateFit1b, band = 2,
       se = FALSE, type = "diffRisk", short = TRUE, quantile = TRUE)

## confidence intervals / p-values computed using 1000 boostrap samples
## (argument se = TRUE and B = 1000) 
ateFit1c <- ate(fit, data = dtS, treatment = "X1",
                times = 5:8, se = TRUE, B = 50, handler = "mclapply")
## NOTE: for real applications 50 bootstrap samples is not enough 

## same but using 2 cpus for generating and analyzing the boostrap samples
## (parallel computation, argument mc.cores = 2) 
ateFit1d <- ate(fit, data = dtS, treatment = "X1",
                times = 5:8, se = TRUE, B = 50, mc.cores = 2)

## manually defining the cluster to be used
## useful when specific packages need to be loaded in each cluster
fit <- cph(formula = Surv(time,event)~ X1+X2+rcs(X6),data=dtS,y=TRUE,x=TRUE)

cl <- parallel::makeCluster(2)
parallel::clusterEvalQ(cl, library(rms))

ateFit1e <- ate(fit, data = dtS, treatment = "X1",
                times = 5:8, se = TRUE, B = 50,
                handler = "foreach", cl = cl)
}

#### Survival settings without censoring ####
#### ATE with glm                        ####

## generate data
n <- 100
dtB <- sampleData(n, outcome="binary")
dtB[, X2 := as.numeric(X2)]
#>            X6       X7          X8           X9          X10 X1 X2 X3 X4 X5 Y
#>   1: 78.01298 63.96298 -0.56839002 -0.210572124  0.502907735  0  2  1  1  1 1
#>   2: 55.52145 54.36513 -0.39144754  0.937064087 -1.374164637  1  1  0  1  1 1
#>   3: 43.13798 62.83007  0.64288621 -0.587024900  0.036800035  0  1  0  1  1 0
#>   4: 49.55549 58.65156  0.27620613  0.496585989  0.586200913  1  1  1  1  0 1
#>   5: 79.22190 60.16542 -0.25813305 -0.260201777  0.683943511  0  1  0  0  1 1
#>   6: 82.94772 62.76167  1.46582983  1.382226397 -0.568756129  0  1  0  0  1 1
#>   7: 38.24775 54.29504 -1.97327590  0.566921974  0.270355642  0  1  1  1  1 0
#>   8: 75.16737 65.93744  0.92010988  0.074845128 -0.541936997  0  1  0  1  1 0
#>   9: 45.05561 61.47713 -0.14907135 -0.910207520  0.402609684  1  1  0  1  1 1
#>  10: 42.73213 56.04234  1.23738748 -0.957253179 -0.186589923  0  1  0  1  1 1
#>  11: 77.12665 53.30601  0.52738054 -1.278404907  0.642750763  0  1  0  1  0 1
#>  12: 58.88072 69.29504  1.49049279  0.007014031  0.671474957  0  1  0  0  1 1
#>  13: 42.79600 66.46572 -0.39243138 -0.180816188 -0.401039690  1  1  0  0  0 1
#>  14: 76.82265 59.70732  0.18796458 -1.139473919  2.165025815  0  1  1  0  0 1
#>  15: 38.35310 65.01524 -0.01127341  0.743817244 -1.395293526  0  1  1  0  1 0
#>  16: 79.90491 66.60928 -1.24592590  0.624441674 -0.170700120  1  1  0  1  0 1
#>  17: 78.05356 68.87915  0.94396826 -1.418233056 -0.169434430  1  2  0  0  1 1
#>  18: 57.61872 64.74185  0.45941533  0.933141193 -0.032983385  0  2  1  0  0 0
#>  19: 66.98958 56.33181 -1.65322897  1.150647382  2.388519770  0  1  1  0  1 0
#>  20: 45.24455 60.60105 -0.77341027  0.470029213 -0.990011932  0  2  0  0  0 0
#>  21: 62.02366 59.91564 -0.68975150  0.879631239 -0.101054172  0  1  0  0  1 1
#>  22: 57.78682 67.08263  1.67183286 -0.923246979 -1.914149078  0  1  0  0  1 1
#>  23: 63.99209 57.24782 -0.09604987  0.251379345 -0.548874372  0  2  1  0  0 1
#>  24: 53.00781 49.67357 -0.79600441 -1.253740771  0.713894914  0  1  0  0  0 1
#>  25: 40.93517 48.01423 -0.22474377 -0.155392632  1.874488735  0  1  1  1  0 0
#>  26: 76.26733 55.95622 -1.26017269  0.517320940  0.794657504  0  1  0  1  1 0
#>  27: 85.67645 65.05618 -1.42640232  0.423064288 -0.094713679  0  1  0  0  1 0
#>  28: 76.45583 53.01383 -2.65363714  1.158858514 -1.347056080  0  2  1  0  0 1
#>  29: 41.69516 68.75821 -1.03865936  1.907944043  0.226424515  0  1  0  1  0 0
#>  30: 68.97958 60.85126  0.57982435  0.122316551  0.024109298  0  1  0  1  1 0
#>  31: 29.92992 56.16979  2.24926361  0.952111614 -1.804701217  0  1  0  0  1 0
#>  32: 55.66421 65.67026 -0.34730097  1.381241504  0.740596491  0  1  0  0  0 1
#>  33: 55.93535 62.49756  0.35892133 -0.688524922 -0.365777394  0  2  1  0  1 0
#>  34: 82.16710 58.59791 -0.56234113 -0.004725565  0.938043750  1  2  0  0  0 1
#>  35: 46.93589 60.39387 -0.96005031  0.706906654 -0.264641308  0  1  0  0  0 0
#>  36: 62.54992 69.20581  0.34253659  0.638144724 -0.166362498  0  1  0  0  1 0
#>  37: 74.74389 59.98933 -1.04648803 -1.787473323 -0.003882510  0  2  0  1  0 1
#>  38: 62.83938 64.19338  0.55333508 -0.205438562  0.435170090  0  1  0  0  1 1
#>  39: 83.37981 69.24409 -0.28051870  0.157280300 -0.620020411  0  1  0  1  0 1
#>  40: 63.87927 56.44113 -0.25400123  0.778093954 -0.245762459  0  2  1  0  0 0
#>  41: 82.95381 63.89724  0.99124934  0.445792808 -0.111607545  0  1  0  1  1 1
#>  42: 76.77325 55.69664 -0.66035357 -0.067550191  0.520877462  1  1  0  1  0 1
#>  43: 40.11926 49.80686  0.69329156 -1.480497207 -0.258175659  0  2  1  1  1 1
#>  44: 62.04355 51.48902  1.52728405 -0.207528914 -0.071121570  0  1  0  1  1 1
#>  45: 38.09094 67.44162  0.23695659 -0.319732621  1.149467397  1  1  1  0  1 1
#>  46: 54.80713 59.21902  1.17387684  1.289086796 -2.430072886  0  1  0  1  1 0
#>  47: 75.85677 61.97361 -2.48614865  0.072783058 -1.710448011  1  1  0  0  0 1
#>  48: 54.50966 70.09970  1.49742470 -0.005325330  0.258522399  0  1  1  1  0 0
#>  49: 74.42121 56.72927  0.97382087  0.818791435  0.498021292  0  1  0  0  1 1
#>  50: 46.25332 62.31992 -0.48411549  0.952465099 -0.049396092  1  1  0  1  0 0
#>  51: 84.39781 61.44562  0.27685067 -1.939096400  1.494676096  0  2  0  0  0 1
#>  52: 66.56635 60.23336 -0.21522285  0.878847068 -2.581771509  0  1  0  0  1 0
#>  53: 47.40846 58.47434  1.70827066 -0.402449468  0.061033497  0  1  0  0  1 1
#>  54: 72.80888 56.05802  1.53843123 -0.882386965  0.327749251  1  1  0  1  1 1
#>  55: 70.58864 60.25092 -0.23621370  0.612407673  0.128658740  0  2  1  0  1 0
#>  56: 66.74386 58.00998 -2.07643183  1.130871707 -0.496879832  0  1  1  0  0 0
#>  57: 68.62644 57.81581 -0.60569297 -1.636657486  0.432174401  0  1  0  0  1 1
#>  58: 87.17309 59.33232  0.95235963  1.083596256  1.412644391  0  1  0  0  0 1
#>  59: 64.16984 59.67586 -1.18769287 -0.085527482 -2.073420227  1  1  0  0  1 1
#>  60: 55.56214 62.99941  3.81258034  0.563125323  0.689403226  0  1  0  0  0 1
#>  61: 80.09402 56.00253  0.64648939  0.372485893  0.653175195  0  1  1  1  1 1
#>  62: 77.64821 51.11732  0.14886884  1.192991625  0.198049842  0  1  1  1  1 0
#>  63: 48.23083 52.25507 -1.07258069  0.734835251  1.399588371  0  1  1  0  1 1
#>  64: 47.79521 66.95377  0.17764396 -1.007523153  0.364184209  0  1  1  0  1 0
#>  65: 79.24335 65.84558 -0.28467953  0.887392638  1.640771446  0  2  1  1  1 1
#>  66: 71.57846 53.46988 -0.70952919  0.418746896  1.135271439  0  1  0  0  1 1
#>  67: 63.67749 67.78430 -0.72730972  0.298064973 -1.216440410  0  1  0  1  1 1
#>  68: 45.57033 64.87571  2.18399129  0.108168399  0.900987013  0  1  1  1  0 1
#>  69: 48.95988 54.12062 -0.42034232  0.040065443 -0.611960104  0  1  0  0  0 1
#>  70: 75.36952 59.78862 -0.13736276  1.471527338  1.131413771  0  2  0  0  1 1
#>  71: 68.28876 59.56319 -0.52263565 -1.755814448 -1.156178361  1  1  0  0  1 1
#>  72: 49.55450 72.63081  1.46846228 -0.092613881 -1.101728513  1  1  0  0  0 1
#>  73: 65.87922 58.28823  1.58985009  1.237809270  0.577847393  0  1  0  1  1 1
#>  74: 74.49191 55.46585 -2.05805583  0.420395072  0.180873659  1  1  0  0  1 1
#>  75: 75.77679 54.73128  1.01236094  0.165181742 -1.987922886  0  1  1  0  1 1
#>  76: 69.50403 51.95586 -1.34335639 -0.827838810  1.360595187  0  1  1  1  0 0
#>  77: 71.86100 61.13832 -1.76114717  0.205799319 -1.174200340  0  1  0  0  1 1
#>  78: 58.22115 57.86385  1.36326293 -0.998953867 -0.937909463  0  1  0  0  0 1
#>  79: 89.64008 61.67966 -0.16707925  0.028508302  1.268314046  0  1  1  1  0 1
#>  80: 57.88654 67.50344 -0.70818084  0.721889354  1.364005221  0  1  0  0  1 0
#>  81: 76.41378 57.91319  0.28394300  0.203508820 -1.842217474  0  1  1  1  0 1
#>  82: 45.59080 63.00873 -0.20854356 -0.249387100 -0.564817488  0  1  0  1  0 0
#>  83: 85.35118 57.55862  0.04651940  2.071403671 -1.659969104  0  1  1  1  1 1
#>  84: 47.72356 63.82289 -1.22547324 -0.982768991  1.252569834  0  2  0  1  0 0
#>  85: 57.45616 55.32343  0.49211534  0.242662434 -1.632399936  0  1  0  0  0 1
#>  86: 57.45603 62.65382  2.22979453  0.003371298  1.053663270  0  1  0  0  1 0
#>  87: 31.38942 60.23798  2.85804162 -1.551135977 -2.768705694  0  1  0  0  0 0
#>  88: 57.83100 60.00459  0.99843840  1.222690877 -1.781053425  0  1  1  1  0 0
#>  89: 65.89807 54.67264 -0.06623606 -0.374327714  0.256339139  0  1  1  0  1 1
#>  90: 49.53415 51.54923 -1.54889847  0.032008937  0.005456635  1  1  1  1  0 1
#>  91: 64.49913 69.79595  0.32266619 -0.230910497  0.741769953  0  1  0  1  1 0
#>  92: 40.62557 62.61934  0.25213485 -0.277975908 -2.685602399  0  1  0  0  0 1
#>  93: 64.88648 59.03623  0.69127232 -1.339535295  1.101621330  0  1  0  1  1 1
#>  94: 66.47275 64.78869  1.18338315 -0.834602773  1.218020447  0  1  0  1  1 1
#>  95: 51.23077 60.85499 -1.53210388  0.331266844  1.501020579  0  1  0  1  0 0
#>  96: 76.13213 51.24694  1.32356115 -0.861312945 -1.185903183  0  1  0  1  0 1
#>  97: 80.60242 65.63773  0.81922078  0.441595794 -1.421862105  0  1  1  0  1 1
#>  98: 67.69269 50.80752  1.80025247  0.730770391 -0.487275186  1  2  0  0  1 1
#>  99: 54.35984 57.30480 -1.61885892  1.378323304  0.589980565  0  1  0  0  1 0
#> 100: 82.50348 56.96561 -0.01577496  0.481238079 -0.454240463  0  1  0  1  0 1
#>            X6       X7          X8           X9          X10 X1 X2 X3 X4 X5 Y

## estimate a logistic regression model
fit <- glm(formula = Y ~ X1+X2, data=dtB, family = "binomial")

## compute the ATE using X1 as the treatment variable
## only point estimate (argument se = FALSE)
ateFit1a <- ate(fit, data = dtB, treatment = "X1", se = FALSE)
#>  Input variables 
#>  - Treatment            : X1 (2 levels: "0" "1")
#>  - Event                : Y (cause: 1, competing risk(s): 0)
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#> 
#>  Processing
#>  - Point estimation: done 
ateFit1a
#>      Average treatment effect for cause 1 
#> 
#>  - Treatment            : X1 (2 levels: "0" "1")
#>  - Event                : Y (cause: 1, competing risk(s): 0)
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#> 
#>  Results 
#>  - Standardized risks   :                
#>  X1=A X1=B    risk.A    risk.B difference (B-A) ratio (B/A)
#>     0    1 0.5853676 0.9444428        0.3590752    1.613418
#> 
#>  - Computation time     : 0.01162887 secs (point estimate)
#> 

if (FALSE) {
## with confidence intervals
ateFit1b <- ate(fit, data = dtB, treatment = "X1",
               times = 5) ## just for having a nice output not used in computations
summary(ateFit1b, short = TRUE)

## using the lava package
library(lava)
ateLava <- estimate(fit, function(p, data){
a <- p["(Intercept)"] ; b <- p["X11"] ; c <- p["X2"] ;
R.X11 <- expit(a + b + c * data[["X2"]])
R.X10 <- expit(a + c * data[["X2"]])
list(risk0=R.X10,risk1=R.X11,riskdiff=R.X11-R.X10)},
average=TRUE)
ateLava
}

## see wglm for handling right-censoring with glm

#### Competing risks settings               ####
#### ATE with cause specific Cox regression ####

## generate data
n <- 500
set.seed(10)
dt <- sampleData(n, outcome="competing.risks")
dt$X1 <- factor(rbinom(n, prob = c(0.2,0.3) , size = 2), labels = paste0("T",0:2))

## estimate cause specific Cox model
fitCR <-  CSC(Hist(time,event)~ X1+X8,data=dt,cause=1)

## compute the ATE at times 1, 5, 10 using X1 as the treatment variable
ateFit2a <- ate(fitCR, data = dt, treatment = "X1", times = c(1,5,10),
                cause = 1, se = TRUE, band = TRUE)
#>  Input variables 
#>  - Treatment            : X1 (3 levels: "T0" "T1" "T2")
#>  - Event                : event (cause: 1, competing risk(s): 2, censoring: 0)
#>  - Time  [min;max]      : time [0.0839;19.1]
#>  - Eval. time           :   1   5 10
#>       number at risk T0   263 106 12
#>       number at risk T1   160  57 11
#>       number at risk T2    26  15  3
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#>  - Uncertainty: Gaussian approximation 
#>                 where the variance is estimated via the influence function 
#> 
#>  Processing
#>  - Prepare influence function: outcome done 
#>  - Point estimation: done 
#>  - Decomposition iid: done
#>  - Confidence intervals / bands: done
summary(ateFit2a)
#>      Average treatment effect for cause 1 
#> 
#>  - Treatment            : X1 (3 levels: "T0" "T1" "T2")
#>  - Event                : event (cause: 1, competing risk(s): 2, censoring: 0)
#>  - Time  [min;max]      : time [0.0839;19.1]
#>  - Eval. time           :   1   5 10
#>       number at risk T0   263 106 12
#>       number at risk T1   160  57 11
#>       number at risk T2    26  15  3
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#>  - Uncertainty: Gaussian approximation 
#>                 where the variance is estimated via the influence function 
#> 
#>  Testing procedure
#>  - Null hypothesis     : given two treatments (A,B), equal risks at all timepoints 
#>  - Confidence level    : 0.95
#>  - Multiple comparisons: single-step max-T adjustment computed using 10000 simulations
#> 
#>  Results: 
#>  - Difference in standardized risk (B-A) between time zero and 'time' 
#>                 reported on the scale [-1;1] (difference between two probabilities)
#>  (difference in average risks when treating all subjects with the experimental treatment (B),
#>                                 vs. treating all subjects with the reference treatment (A))
#> 
#>  time X1=A risk(X1=A) X1=B risk(X1=B) difference           ci p.value
#>     1   T0     0.0673   T1     0.0681   8.72e-04 [-0.02;0.02]   0.926
#>     5   T0     0.3744   T1     0.3743  -2.52e-05 [-0.08;0.08]   1.000
#>    10   T0     0.5017   T1     0.4935  -8.21e-03 [-0.10;0.09]   0.867
#>     1   T0     0.0673   T2     0.0523  -1.49e-02 [-0.05;0.02]   0.355
#>     5   T0     0.3744   T2     0.3063  -6.81e-02 [-0.22;0.09]   0.388
#>    10   T0     0.5017   T2     0.4253  -7.64e-02 [-0.27;0.12]   0.434
#>     1   T1     0.0681   T2     0.0523  -1.58e-02 [-0.05;0.02]   0.346
#>     5   T1     0.3743   T2     0.3063  -6.80e-02 [-0.23;0.09]   0.404
#>    10   T1     0.4935   T2     0.4253  -6.82e-02 [-0.27;0.13]   0.498
#>  simultaneous ci adjusted p.value
#>     [-0.02;0.02]            0.987
#>     [-0.09;0.09]            1.000
#>     [-0.11;0.09]            0.953
#>     [-0.05;0.02]            0.419
#>     [-0.23;0.10]            0.454
#>     [-0.28;0.13]            0.507
#>     [-0.05;0.02]            0.409
#>     [-0.24;0.10]            0.472
#>     [-0.28;0.14]            0.581
#> 
#>  difference      : estimated difference in standardized risks 
#>  ci              : pointwise confidence intervals 
#>  p.value         : (unadjusted) p-value 
#>  simultaneous ci : simulatenous confidence intervals over time
#>  adjusted p.value: p-value adjusted for multiple comparisons over time
as.data.table(ateFit2a)
#>          type estimator time level      estimate          se       lower
#>  1:  meanRisk  GFORMULA    1    T0  6.725075e-02 0.012050231  0.04363273
#>  2:  meanRisk  GFORMULA    5    T0  3.743564e-01 0.027863621  0.31974466
#>  3:  meanRisk  GFORMULA   10    T0  5.017158e-01 0.032588024  0.43784440
#>  4:  meanRisk  GFORMULA    1    T1  6.812318e-02 0.012567904  0.04349054
#>  5:  meanRisk  GFORMULA    5    T1  3.743311e-01 0.034694830  0.30633052
#>  6:  meanRisk  GFORMULA   10    T1  4.935039e-01 0.039653201  0.41578504
#>  7:  meanRisk  GFORMULA    1    T2  5.232115e-02 0.018006403  0.01702925
#>  8:  meanRisk  GFORMULA    5    T2  3.062972e-01 0.075452662  0.15841267
#>  9:  meanRisk  GFORMULA   10    T2  4.253305e-01 0.093198889  0.24266407
#> 10:  diffRisk  GFORMULA    1 T0.T1  8.724303e-04 0.009441589 -0.01763274
#> 11:  diffRisk  GFORMULA    5 T0.T1 -2.520916e-05 0.041528106 -0.08141880
#> 12:  diffRisk  GFORMULA   10 T0.T1 -8.211865e-03 0.049084017 -0.10441477
#> 13:  diffRisk  GFORMULA    1 T0.T2 -1.492961e-02 0.016145110 -0.04657344
#> 14:  diffRisk  GFORMULA    5 T0.T2 -6.805918e-02 0.078813390 -0.22253059
#> 15:  diffRisk  GFORMULA   10 T0.T2 -7.638522e-02 0.097706198 -0.26788585
#> 16:  diffRisk  GFORMULA    1 T1.T2 -1.580204e-02 0.016761368 -0.04865371
#> 17:  diffRisk  GFORMULA    5 T1.T2 -6.803397e-02 0.081442901 -0.22765912
#> 18:  diffRisk  GFORMULA   10 T1.T2 -6.817335e-02 0.100625523 -0.26539576
#> 19: ratioRisk  GFORMULA    1 T0.T1  1.012973e+00 0.141191244  0.73624304
#> 20: ratioRisk  GFORMULA    5 T0.T1  9.999327e-01 0.110929211  0.78251540
#> 21: ratioRisk  GFORMULA   10 T0.T1  9.836324e-01 0.097205064  0.79311401
#> 22: ratioRisk  GFORMULA    1 T0.T2  7.780009e-01 0.234968184  0.31747172
#> 23: ratioRisk  GFORMULA    5 T0.T2  8.181968e-01 0.206945429  0.41259122
#> 24: ratioRisk  GFORMULA   10 T0.T2  8.477520e-01 0.191990382  0.47145777
#> 25: ratioRisk  GFORMULA    1 T1.T2  7.680373e-01 0.237987611  0.30159016
#> 26: ratioRisk  GFORMULA    5 T1.T2  8.182519e-01 0.211754010  0.40322167
#> 27: ratioRisk  GFORMULA   10 T1.T2  8.618585e-01 0.199975344  0.46991406
#>          type estimator time level      estimate          se       lower
#>          upper quantileBand   lowerBand  upperBand   p.value adj.p.value
#>  1: 0.09086877     2.317771  0.03932108 0.09518043        NA          NA
#>  2: 0.42896804     2.317771  0.30977486 0.43893784        NA          NA
#>  3: 0.56558711     2.317771  0.42618418 0.57724733        NA          NA
#>  4: 0.09275582     2.300197  0.03921453 0.09703183        NA          NA
#>  5: 0.44233176     2.300197  0.29452621 0.45413608        NA          NA
#>  6: 0.57122274     2.300197  0.40229373 0.58471405        NA          NA
#>  7: 0.08761305     2.173021  0.01319285 0.09144944        NA          NA
#>  8: 0.45418167     2.173021  0.14233695 0.47025739        NA          NA
#>  9: 0.60799700     2.173021  0.22280739 0.62785368        NA          NA
#> 10: 0.01937760     2.077331 -0.01874087 0.02048573 0.9263779      0.9866
#> 11: 0.08136838     2.077331 -0.08629282 0.08624240 0.9995157      1.0000
#> 12: 0.08799104     2.077331 -0.11017560 0.09375187 0.8671323      0.9528
#> 13: 0.01671423     2.100121 -0.04883630 0.01897708 0.3551148      0.4188
#> 14: 0.08641222     2.100121 -0.23357685 0.09745849 0.3878360      0.4544
#> 15: 0.11511541     2.100121 -0.28158008 0.12880964 0.4343411      0.5071
#> 16: 0.01704964     2.101442 -0.05102508 0.01942101 0.3458009      0.4095
#> 17: 0.09159118     2.101442 -0.23918153 0.10311359 0.4035162      0.4725
#> 18: 0.12904905     2.101442 -0.27963209 0.14328538 0.4980915      0.5807
#> 19: 1.28970255     2.090655  0.71779060 1.30815499 0.9267926      0.9851
#> 20: 1.21734992     2.090655  0.76801794 1.23184738 0.9995156      1.0000
#> 21: 1.17415086     2.090655  0.78041017 1.18685470 0.8662829      0.9563
#> 22: 1.23853008     2.011050  0.30546805 1.25053374 0.3447585      0.3916
#> 23: 1.22380239     2.011050  0.40201913 1.23437448 0.3796681      0.4292
#> 24: 1.22404624     2.011050  0.46164968 1.23385433 0.4277790      0.4786
#> 25: 1.23448445     2.024729  0.28617680 1.24989782 0.3297171      0.3747
#> 26: 1.23328214     2.024729  0.38950734 1.24699647 0.3907278      0.4408
#> 27: 1.25380300     2.024729  0.45696257 1.26675448 0.4896959      0.5518
#>          upper quantileBand   lowerBand  upperBand   p.value adj.p.value

#### Double robust estimator ####
if (FALSE) {
## generate data
n <- 500
set.seed(10)
dt <- sampleData(n, outcome="competing.risks")
dt$time <- round(dt$time,1)
dt$X1 <- factor(rbinom(n, prob = c(0.4) , size = 1), labels = paste0("T",0:1))

## working models
m.event <-  CSC(Hist(time,event)~ X1+X2+X3+X5+X8,data=dt)
m.censor <-  coxph(Surv(time,event==0)~ X1+X2+X3+X5+X8,data=dt, x = TRUE, y = TRUE)
m.treatment <-  glm(X1~X2+X3+X5+X8,data=dt,family=binomial(link="logit"))

## prediction + average
ateRobust <- ate(event = m.event,
                 treatment = m.treatment,
                 censor = m.censor,
                 data = dt, times = 5:10, 
                 cause = 1, band = TRUE)

## compare various estimators
ateRobust3 <- ate(event = m.event,
                 treatment = m.treatment,
                 censor = m.censor,
                 estimator = c("GFORMULA","IPTW","AIPTW"),
                 data = dt, times = c(5:10), 
                 cause = 1, se = TRUE)
print(setkeyv(as.data.table(ateRobust3, type = "meanRisk"),"time"))
print(setkeyv(as.data.table(ateRobust3, type = "diffRisk"),"time"))
}

#### time-dependent covariates ###
if (FALSE) {
library(survival)
fit <- coxph(Surv(time, status) ~ celltype+karno + age + trt, veteran)
vet2 <- survSplit(Surv(time, status) ~., veteran,
                       cut=c(60, 120), episode ="timegroup")
fitTD <- coxph(Surv(tstart, time, status) ~ celltype +karno + age + trt,
               data= vet2,x=1)
set.seed(16)
resVet <- ate(fitTD,formula=Hist(entry=tstart,time=time,event=status)~1,
          data = vet2, treatment = "celltype",
        times=5,verbose=1,
        landmark = c(0,30,60,90), cause = 1, B = 50, se = 1,
        band = FALSE, mc.cores=1)
summary(resVet)
}

if (FALSE) {
set.seed(137)
d=sampleDataTD(127)
library(survival)
d[,status:=1*(event==1)]
d[,X3:=as.factor(X3)]
## ignore competing risks
cox1TD <- coxph(Surv(start,time, status,type="counting") ~ X3+X5+X6+X8,
                data=d, x = TRUE)
resTD1 <- ate(cox1TD,formula=Hist(entry=start,time=time,event=status)~1,
        data = d, treatment = "X3", contrasts = NULL,
        times=.5,verbose=1,
        landmark = c(0,0.5,1), B = 20, se = 1,
        band = FALSE, mc.cores=1)
resTD1
## account for competing risks
cscTD <- CSC(Hist(time=time, event=event,entry=start) ~ X3+X5+X6+X8, data=d)
set.seed(16)
resTD <- ate(cscTD,formula=Hist(entry=start,time=time,event=event)~1,
        data = d, treatment = "X3", contrasts = NULL,
        times=.5,verbose=1,
        landmark = c(0,0.5,1), cause = 1, B = 20, se = 1,
        band = FALSE, mc.cores=1)
resTD
}