Standard error of the absolute risk predicted from cause-specific Cox models

Standard error of the absolute risk predicted from cause-specific Cox models using a first order von Mises expansion of the absolute risk functional.

Usage

calcSeCSC(
  object,
  cif,
  hazard,
  cumhazard,
  survival,
  object.time,
  object.maxtime,
  eXb,
  new.LPdata,
  new.strata,
  times,
  surv.type,
  ls.infoVar,
  new.n,
  cause,
  nCause,
  nVar.lp,
  export,
  store.iid,
  diag
)

Arguments

object

The fitted cause specific Cox model

cif

the cumulative incidence function at each prediction time for each individual.

hazard

list containing the baseline hazard for each cause in a matrix form. Columns correspond to the strata.

cumhazard

list containing the cumulative baseline hazard for each cause in a matrix form. Columns correspond to the strata.

survival

list containing the (all cause) survival in a matrix form at t-. Columns correspond to event times.

object.time

a vector containing all the events regardless to the cause.

object.maxtime

a matrix containing the latest event in the strata of the observation for each cause.

eXb

a matrix containing the exponential of the linear predictor evaluated for the new observations (rows) for each cause (columns)

new.LPdata

a list of design matrices for the new observations for each cause.

new.strata

a matrix containing the strata indicator for each observation and each cause.

times

the time points at which to evaluate the predictions.

surv.type

see the surv.type argument of CSC.

ls.infoVar

A list containing the output of coxVariableName for each Cox model.

new.n

the number of new observations.

cause

the cause of interest.

nCause

the number of causes.

nVar.lp

the number of variables that form the linear predictor in each Cox model

export

can be "iid" to return the value of the influence function for each observation "se" to return the standard error for a given timepoint

store.iid

the method used to compute the influence function and the standard error. Can be "full" or "minimal". See the details section.

diag

[logical] when FALSE the absolute risk/survival for all observations at all times is computed, otherwise it is only computed for the i-th observation at the i-th time.

Details

Can also return the empirical influence function of the functionals cumulative hazard or survival or the sum over the observations of the empirical influence function.

store.iid="full" compute the influence function for each observation at each time in the argument times before computing the standard error / influence functions. store.iid="minimal" recompute for each subject specific prediction the influence function for the baseline hazard. This avoid to store all the influence functions but may lead to repeated evaluation of the influence function. This solution is therefore efficient more efficient in memory usage but may not be in term of computation time.