Computation of standard errors for predictions

Compute the standard error associated to the predictions from Cox regression model using a first order von Mises expansion of the functional (cumulative hazard or survival).

Usage

calcSeCox(
  object,
  times,
  nTimes,
  type,
  diag,
  Lambda0,
  object.n,
  object.time,
  object.eXb,
  object.strata,
  nStrata,
  new.n,
  new.eXb,
  new.LPdata,
  new.strata,
  new.survival,
  nVar.lp,
  export,
  store.iid
)

Arguments

object

The fitted Cox regression model object either obtained with coxph (survival package) or cph (rms package).

times

Vector of times at which to return the estimated hazard/survival.

nTimes

the length of the argument times.

type

One or several strings that match (either in lower or upper case or mixtures) one or several of the strings "hazard","cumhazard", "survival".

diag

[logical] when FALSE the hazard/cumlative hazard/survival for all observations at all times is computed, otherwise it is only computed for the i-th observation at the i-th time.

Lambda0

the baseline hazard estimate returned by BaseHazStrata_cpp.

object.n

the number of observations in the dataset used to estimate the object.

object.time

the time to event of the observations used to estimate the object.

object.eXb

the exponential of the linear predictor relative to the observations used to estimate the object.

object.strata

the strata index of the observations used to estimate the object.

nStrata

the number of strata.

new.n

the number of observations for which the prediction was performed.

new.eXb

the linear predictor evaluated for the new observations.

new.LPdata

the variables involved in the linear predictor for the new observations.

new.strata

the strata indicator for the new observations.

new.survival

the survival evaluated for the new observations.

nVar.lp

the number of variables that form the linear predictor.

export

can be "iid" to return the value of the influence function for each observation. "se" to return the standard error for a given timepoint. "average.iid" to return the value of the average influence function over the observations for which the prediction was performed.

store.iid

Implementation used to estimate the influence function and the standard error. Can be "full" or "minimal". See the details section.

Value

A list optionally containing the standard error for the survival, cumulative hazard and hazard.

Details

store.iid="full" compute the influence function for each observation at each time in the argument times before computing the standard error / influence functions. store.iid="minimal" recompute for each subject specific prediction the influence function for the baseline hazard. This avoid to store all the influence functions but may lead to repeated evaluation of the influence function. This solution is therefore more efficient in memory usage but may not be in terms of computation time.

Author

Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk