Confidence Intervals and Confidence Bands for the Predicted Absolute Risk (Cumulative Incidence Function)

Confidence intervals and confidence Bands for the predicted absolute risk (cumulative incidence function).

Usage

# S3 method for ate
confint(
  object,
  parm = NULL,
  level = 0.95,
  n.sim = 10000,
  estimator = object$estimator,
  contrasts = object$contrasts,
  allContrasts = object$allContrasts,
  meanRisk.transform = "none",
  diffRisk.transform = "none",
  ratioRisk.transform = "none",
  seed = NA,
  ci = object$inference$se,
  band = object$inference$band,
  p.value = TRUE,
  method.band = "maxT-simulation",
  alternative = "two.sided",
  bootci.method = "perc",
  ...
)

Arguments

object

A ate object, i.e. output of the ate function.

parm

not used. For compatibility with the generic method.

level

[numeric, 0-1] Level of confidence.

n.sim

[integer, >0] the number of simulations used to compute the quantiles for the confidence bands and/or perform adjustment for multiple comparisons.

estimator

[character] The type of estimator relative to which the estimates should be displayed.

contrasts

[character vector] levels of the treatment variable for which the risks should be assessed and compared. Default is to consider all levels.

allContrasts

[2-row character matrix] levels of the treatment variable to be compared. Default is to consider all pairwise comparisons.

meanRisk.transform

[character] the transformation used to improve coverage of the confidence intervals for the mean risk in small samples. Can be "none", "log", "loglog", "cloglog".

diffRisk.transform

[character] the transformation used to improve coverage of the confidence intervals for the risk difference in small samples. Can be "none", "atanh".

ratioRisk.transform

[character] the transformation used to improve coverage of the confidence intervals for the risk ratio in small samples. Can be "none", "log".

seed

[integer, >0] seed number set when performing simulation for the confidence bands. If not given or NA no seed is set.

ci

[logical] should the confidence intervals be computed?

band

[logical] should the confidence bands be computed?

p.value

[logical] should the p-values/adjusted p-values be computed? Requires argument ci and/or band to be TRUE.

method.band

[character] method used to adjust for multiple comparisons. Can be any element of p.adjust.methods (e.g. "holm"), "maxT-integration", or "maxT-simulation".

alternative

[character] a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".

bootci.method

[character] Method for constructing bootstrap confidence intervals. Either "perc" (the default), "norm", "basic", "stud", or "bca".

...

not used.

Details

Argument ci, band, p.value, method.band, alternative, meanRisk.transform, diffRisk.transform, ratioRisk.transform are only active when the ate object contains the influence function. Argument bootci.method is only active when the ate object contains bootstrap samples.

Influence function: confidence bands and confidence intervals computed via the influence function are automatically restricted to the interval of definition of the parameter (e.g. [0;1] for the average risk). Single step max adjustment for multiple comparisons, i.e. accounting for the correlation between the test statistics but not for the ordering of the tests, can be performed setting the arguemnt method.band to "maxT-integration" or "maxT-simulation". The former uses numerical integration (pmvnorm and qmvnorm to perform the adjustment while the latter using simulation. Both assume that the test statistics are jointly normally distributed.

Bootstrap: confidence intervals obtained via bootstrap are computed using the boot.ci function of the boot package. p-value are obtained using test inversion method (finding the smallest confidence level such that the interval contain the null hypothesis).

Author

Brice Ozenne

Examples

library(survival)
library(data.table)

## ## generate data ####
set.seed(10)
d <- sampleData(70,outcome="survival")
d[, X1 := paste0("T",rbinom(.N, size = 2, prob = c(0.51)))]
#>           X6       X7           X8          X9         X10 X1 X2 X3 X4 X5
#>  1: 73.81825 62.37447 -0.886787966  0.32751014  0.14188028 T2  0  1  1  0
#>  2: 55.03195 66.16529  1.271146001  0.06386348  1.26171505 T2  1  0  0  1
#>  3: 79.30851 57.03436 -1.605085441 -1.13956094 -0.43150033 T1  0  1  0  1
#>  4: 60.85891 63.76131  1.122273383  1.18041020 -1.82271259 T0  1  0  1  1
#>  5: 26.94100 63.06023  2.158438642  0.04138848  0.35254396 T0  0  0  0  0
#>  6: 52.15634 58.85671  0.428246613 -1.21359356 -1.34845144 T1  1  1  1  1
#>  7: 76.21178 56.47580  1.201178690  0.07319578  0.70768832 T1  1  1  0  0
#>  8: 56.35288 66.14758  1.031690116 -0.25732459 -0.41089094 T0  0  0  1  1
#>  9: 46.35830 67.59258  0.653874253  0.26680641 -0.44604518 T0  0  1  0  0
#> 10: 46.97551 66.72547  2.012081773  1.38772433 -1.04115630 T1  1  0  1  0
#> 11: 73.04213 59.38154  1.094629842  0.19307964 -0.32922472 T2  1  0  1  1
#> 12: 49.79986 64.66951  0.726409354  0.59231662 -0.28282162 T1  1  0  0  0
#> 13: 62.59822 60.09748 -1.399657137 -0.82997453  0.43242913 T1  0  1  0  0
#> 14: 57.60843 64.15426 -1.287795163  0.39257328 -0.30760710 T1  1  1  0  1
#> 15: 71.90249 58.47753 -0.968415478  0.38486761 -0.05663631 T2  0  0  0  0
#> 16: 85.41526 59.70359 -1.736685758  1.05104467  0.73351542 T1  0  0  0  0
#> 17: 78.59953 65.03698  0.831697358  1.15579748  0.09731162 T1  1  0  0  1
#> 18: 60.44155 53.11102 -0.758953109 -1.03443796  1.63089174 T2  0  1  0  0
#> 19: 84.85822 54.97899  2.155077001 -0.25446807  0.56061070 T0  0  1  1  0
#> 20: 76.96835 59.88360 -0.423258880  1.27368427  1.32956476 T1  0  0  0  0
#> 21: 38.96384 67.73374 -0.005649937  1.50254464 -0.27882350 T2  0  0  1  0
#> 22: 62.35156 56.20318  0.400361729  0.59040947 -1.26673154 T2  0  0  0  1
#> 23: 73.01263 65.03396 -0.365677802 -0.63068545 -0.24914839 T1  0  1  0  1
#> 24: 59.50147 61.37107 -0.429143835  0.79234954  0.01798841 T2  0  0  0  1
#> 25: 74.31097 59.59809  0.244697127  0.12538457  0.37707273 T2  1  1  1  0
#> 26: 76.95145 55.48224  2.219776249  0.32275497  0.79600856 T2  0  0  0  1
#> 27: 65.90546 62.36573  0.314269808 -0.44558321 -0.84067742 T1  0  1  1  0
#> 28: 65.65212 66.94244 -0.563785075  0.76684393 -2.20547175 T0  0  0  1  1
#> 29: 56.22386 60.17373 -1.485696158 -1.40350302 -1.12805599 T1  0  1  1  1
#> 30: 67.56142 53.53707  0.415298593 -1.17604678 -1.34130996 T2  1  1  0  0
#> 31: 40.05295 65.81984 -0.528063683  0.51159647  1.60511404 T2  0  0  0  0
#> 32: 58.22551 52.38308 -0.350470617  1.31676526  0.74442373 T2  0  0  1  0
#> 33: 50.50127 47.40832  0.730621937  2.39291295  0.86208221 T0  0  0  1  0
#> 34: 70.16537 56.46235 -0.556624165 -0.06776941  0.39515582 T1  0  0  1  0
#> 35: 55.53177 58.56284 -1.449760536 -0.20378507  0.50911870 T1  1  0  1  1
#> 36: 69.41883 57.82346 -1.051996428  1.99643445 -0.12255013 T2  0  1  0  1
#> 37: 85.17646 58.28461 -0.682838619  1.54321779  0.09258465 T2  0  0  0  1
#> 38: 23.99996 59.80345 -0.612941179 -1.22833763 -0.35787991 T2  0  1  0  0
#> 39: 39.75236 64.43483  1.890780035  2.64448799 -0.35965522 T2  0  1  0  1
#> 40: 78.03012 58.20026 -0.145410534  0.75701371  1.02857072 T2  0  1  0  0
#> 41: 79.40764 56.35523 -0.639423721  1.05534519  1.07789259 T2  0  0  0  1
#> 42: 59.10608 68.75873 -0.411176045  0.95794490  0.93178122 T0  0  1  0  1
#> 43: 49.72406 59.56981  0.342791878  0.83387906 -1.46079387 T1  0  0  0  0
#> 44: 38.19841 65.53182 -1.133246755 -0.16927557 -0.90607558 T1  1  0  0  1
#> 45: 79.41190 59.90411  0.066070303 -0.40443364 -0.68034783 T2  0  0  0  1
#> 46: 77.64254 52.53270  0.037999766  0.29560157  1.06316604 T0  0  1  1  1
#> 47: 72.96068 73.50318  0.921064748 -0.23187451 -0.69241315 T0  0  0  1  0
#> 48: 25.54901 56.28153  2.134965638 -0.44807823 -1.13362827 T1  0  0  1  1
#> 49: 39.12285 65.32597 -0.599826932  0.02231080 -1.09461543 T1  0  0  1  1
#> 50: 39.21185 59.83959  1.244446459  0.04446730 -1.01290362 T2  1  1  0  1
#> 51: 43.90430 57.30020 -0.079457017  0.48072117  0.41027733 T1  0  1  0  0
#> 52: 45.87862 50.37178  1.181751550  0.63561986  0.47774044 T1  0  0  0  0
#> 53: 51.20349 65.03916  2.186144063 -2.00074239 -2.32987955 T1  1  1  0  0
#> 54: 75.52082 58.60383  0.406174931 -0.69065550  0.01625547 T2  0  0  0  1
#> 55: 66.88443 53.75217 -0.738359105 -0.13064519  0.98035352 T2  0  0  1  0
#> 56: 42.09832 66.24246 -1.956486703 -0.34425178  0.80634826 T2  0  0  0  0
#> 57: 71.87834 59.03893 -1.950045702 -0.05241282  0.11960089 T1  0  0  0  1
#> 58: 33.41543 64.55988 -0.940997816  0.09612392 -2.43661560 T2  0  0  0  0
#> 59: 51.18416 49.20865  1.197786769  0.26640705  0.03102248 T1  0  0  1  1
#> 60: 40.76445 66.44805 -0.624239390  0.55479348 -0.33574670 T1  1  1  1  1
#> 61: 55.31798 57.85945 -0.132869655  1.23544588 -0.26377616 T0  0  0  0  1
#> 62: 49.51837 58.72499 -0.017220922  0.28513338 -0.35585246 T1  0  1  0  0
#> 63: 20.34225 66.63579 -0.459629282 -0.45879293 -0.48115480 T1  0  0  1  1
#> 64: 72.37534 64.19345  1.472936791  0.62347802  0.22273547 T2  0  0  1  0
#> 65: 64.60389 52.67500  2.169319613 -0.72399509  2.42991032 T2  0  1  0  0
#> 66: 59.12941 56.91874 -3.001431306  1.61941870  1.49629812 T0  0  1  1  0
#> 67: 73.89854 60.53053 -1.771985583 -0.61667341 -0.71720820 T1  1  1  1  1
#> 68: 56.71962 50.77992 -0.364839223  0.38454106 -0.46705390 T1  1  1  1  1
#> 69: 74.89700 51.91789  0.374533781  2.11452131  0.66292345 T1  0  1  0  0
#> 70: 31.14998 61.13838 -1.234077927 -1.02856737  2.30017689 T2  1  0  1  1
#>           X6       X7           X8          X9         X10 X1 X2 X3 X4 X5
#>      eventtime  censtime       time event
#>  1:  9.4184480  8.235906  8.2359063     0
#>  2:  5.4761627 10.870427  5.4761627     1
#>  3:  1.0323078  9.225982  1.0323078     1
#>  4: 19.4574401  6.054568  6.0545684     0
#>  5:  0.5112391 15.695559  0.5112391     1
#>  6:  3.3054768 12.205393  3.3054768     1
#>  7:  1.2161130 11.369665  1.2161130     1
#>  8:  8.9251320 11.405403  8.9251320     1
#>  9: 21.4591920  6.962654  6.9626536     0
#> 10: 23.4557337  9.190943  9.1909426     0
#> 11:  3.2179073  6.543997  3.2179073     1
#> 12: 25.1049438  7.523933  7.5239329     0
#> 13:  1.9147684 14.750842  1.9147684     1
#> 14: 30.2969350  7.194720  7.1947199     0
#> 15:  8.3233198 10.134509  8.3233198     1
#> 16:  6.2367495  9.201862  6.2367495     1
#> 17:  4.9619976 17.199920  4.9619976     1
#> 18:  4.0581326 11.537476  4.0581326     1
#> 19:  0.4985251  9.588110  0.4985251     1
#> 20:  1.7957700  4.230440  1.7957700     1
#> 21: 14.4476473  3.812455  3.8124547     0
#> 22:  3.7826564  6.968215  3.7826564     1
#> 23:  1.1642427  5.047281  1.1642427     1
#> 24:  1.3077129 10.168760  1.3077129     1
#> 25:  1.4479584  9.496166  1.4479584     1
#> 26:  2.7435111  5.892573  2.7435111     1
#> 27:  3.6877980  4.199928  3.6877980     1
#> 28:  8.5142430 11.953367  8.5142430     1
#> 29: 10.9359909  5.102580  5.1025805     0
#> 30:  1.1212223 10.164211  1.1212223     1
#> 31: 46.1189084  7.901729  7.9017289     0
#> 32:  8.6184778 15.408467  8.6184778     1
#> 33:  9.8688062 13.315841  9.8688062     1
#> 34:  7.3795769  3.248795  3.2487951     0
#> 35: 20.8801838  9.280303  9.2803027     0
#> 36: 20.5985926  5.391580  5.3915796     0
#> 37:  8.1520810  4.418386  4.4183855     0
#> 38: 27.5982200  2.154221  2.1542209     0
#> 39: 13.3874120  6.145630  6.1456304     0
#> 40:  4.5201609  8.319503  4.5201609     1
#> 41:  9.1908871 11.354410  9.1908871     1
#> 42: 31.0394522 12.142705 12.1427049     0
#> 43: 23.7193811 20.586829 20.5868289     0
#> 44: 51.2726204  5.622545  5.6225453     0
#> 45:  5.1087674 11.776131  5.1087674     1
#> 46:  4.8693016 13.510045  4.8693016     1
#> 47:  7.0570522 20.179287  7.0570522     1
#> 48:  5.1575500  8.493175  5.1575500     1
#> 49: 49.1905410 15.079774 15.0797744     0
#> 50:  7.2212702  4.703314  4.7033144     0
#> 51: 11.2465692 10.185932 10.1859318     0
#> 52: 26.2233288  2.562768  2.5627679     0
#> 53:  4.0968780 11.844691  4.0968780     1
#> 54:  2.0576730  8.650785  2.0576730     1
#> 55:  2.3751796 12.855153  2.3751796     1
#> 56:  8.8761931  7.342938  7.3429383     0
#> 57:  3.6109236  8.818587  3.6109236     1
#> 58: 35.8238897  8.720648  8.7206480     0
#> 59:  6.7798993  9.574494  6.7798993     1
#> 60: 56.7600799  8.262151  8.2621515     0
#> 61: 22.1808138 18.562114 18.5621136     0
#> 62: 16.7422707 14.722308 14.7223081     0
#> 63: 82.9200663  8.702242  8.7022417     0
#> 64:  7.2858216  9.611724  7.2858216     1
#> 65:  1.2857034  4.265895  1.2857034     1
#> 66: 41.9847051  5.224576  5.2245761     0
#> 67:  3.5260250  7.458201  3.5260250     1
#> 68:  6.7892548  8.966944  6.7892548     1
#> 69:  0.9575123 15.745788  0.9575123     1
#> 70: 15.9073575 12.320922 12.3209217     0
#>      eventtime  censtime       time event
## table(d$X1)

#### stratified Cox model ####
fit <- coxph(Surv(time,event)~X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)

#### average treatment effect ####
fit.ate <- ate(fit, treatment = "X1", times = 1:3, data = d,
               se = TRUE, iid = TRUE, band = TRUE)
#>  Input variables 
#>  - Treatment            : X1 (3 levels: "T0" "T1" "T2")
#>  - Event                : event (cause: 1, censoring: 0)
#>  - Time  [min;max]      : time [0.499;20.6]
#>  - Eval. time           :  1  2  3
#>       number at risk T0   10 10 10
#>       number at risk T1   29 24 23
#>       number at risk T2   28 24 20
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#>  - Uncertainty: Gaussian approximation 
#>                 where the variance is estimated via the influence function 
#> 
#>  Processing
#>  - Prepare influence function: outcome done 
#>  - Point estimation: done 
#>  - Decomposition iid: done
#>  - Confidence intervals / bands: done
summary(fit.ate)
#>      Average treatment effect 
#> 
#>  - Treatment            : X1 (3 levels: "T0" "T1" "T2")
#>  - Event                : event (cause: 1, censoring: 0)
#>  - Time  [min;max]      : time [0.499;20.6]
#>  - Eval. time           :  1  2  3
#>       number at risk T0   10 10 10
#>       number at risk T1   29 24 23
#>       number at risk T2   28 24 20
#> 
#>  Estimation procedure 
#>  - Estimator  : G-formula
#>  - Uncertainty: Gaussian approximation 
#>                 where the variance is estimated via the influence function 
#> 
#>  Testing procedure
#>  - Null hypothesis     : given two treatments (A,B), equal risks at all timepoints 
#>  - Confidence level    : 0.95
#>  - Multiple comparisons: single-step max-T adjustment computed using 10000 simulations
#> 
#>  Results: 
#>  - Difference in standardized risk (B-A) between time zero and 'time' 
#>                 reported on the scale [-1;1] (difference between two probabilities)
#>  (difference in average risks when treating all subjects with the experimental treatment (B),
#>                                 vs. treating all subjects with the reference treatment (A))
#> 
#>  time X1=A risk(X1=A) X1=B risk(X1=B) difference           ci p.value
#>     1   T0     0.0339   T1     0.0472    0.01332 [-0.02;0.05]   0.411
#>     2   T0     0.1352   T1     0.1826    0.04739 [-0.07;0.17]   0.431
#>     3   T0     0.1703   T1     0.2267    0.05642 [-0.08;0.20]   0.432
#>     1   T0     0.0339   T2     0.0420    0.00815 [-0.02;0.03]   0.506
#>     2   T0     0.1352   T2     0.1645    0.02933 [-0.07;0.12]   0.545
#>     3   T0     0.1703   T2     0.2054    0.03511 [-0.08;0.15]   0.551
#>     1   T1     0.0472   T2     0.0420   -0.00518 [-0.04;0.03]   0.764
#>     2   T1     0.1826   T2     0.1645   -0.01806 [-0.13;0.10]   0.759
#>     3   T1     0.2267   T2     0.2054   -0.02131 [-0.16;0.11]   0.758
#>  simultaneous ci adjusted p.value
#>     [-0.02;0.05]            0.504
#>     [-0.08;0.18]            0.525
#>     [-0.10;0.21]            0.526
#>     [-0.02;0.03]            0.590
#>     [-0.07;0.13]            0.630
#>     [-0.09;0.16]            0.636
#>     [-0.04;0.03]            0.811
#>     [-0.14;0.10]            0.806
#>     [-0.16;0.12]            0.805
#> 
#>  difference      : estimated difference in standardized risks 
#>  ci              : pointwise confidence intervals 
#>  p.value         : (unadjusted) p-value 
#>  simultaneous ci : simulatenous confidence intervals over time
#>  adjusted p.value: p-value adjusted for multiple comparisons over time
dt.ate <- as.data.table(fit.ate)

## manual calculation of se
dd <- copy(d)
dd$X1 <- rep(factor("T0", levels = paste0("T",0:2)), NROW(dd))
out <- predictCox(fit, newdata = dd, se = TRUE, times = 1:3, average.iid = TRUE)
term1 <- -out$survival.average.iid
term2 <- sweep(1-out$survival, MARGIN = 2, FUN = "-", STATS = colMeans(1-out$survival))
sqrt(colSums((term1 + term2/NROW(d))^2)) 
#> [1] 0.02557002 0.05576405 0.06256730
## fit.ate$meanRisk[treatment=="T0",se]

## note
out2 <- predictCox(fit, newdata = dd, se = TRUE, times = 1:3, iid = TRUE)
mean(out2$survival.iid[1,1,])
#> [1] -0.0003529296
out$survival.average.iid[1,1]
#> [1] -0.0003529296

## check confidence intervals (no transformation)
dt.ate[,.(lower = pmax(0,estimate + qnorm(0.025) * se),
          lower2 = lower,
          upper = estimate + qnorm(0.975) * se,
          upper2 = upper)]
#>          lower      lower2      upper     upper2
#>  1: 0.00000000  0.00000000 0.08401352 0.08401352
#>  2: 0.02590437  0.02590437 0.24449544 0.24449544
#>  3: 0.04767637  0.04767637 0.29293567 0.29293567
#>  4: 0.00000000  0.00000000 0.10199266 0.10199266
#>  5: 0.06587254  0.06587254 0.29931345 0.29931345
#>  6: 0.09787653  0.09787653 0.35557496 0.35557496
#>  7: 0.00000000  0.00000000 0.08763162 0.08763162
#>  8: 0.07373281  0.07373281 0.25532659 0.25532659
#>  9: 0.10169224  0.10169224 0.30914481 0.30914481
#> 10: 0.00000000 -0.01843023 0.04508014 0.04508014
#> 11: 0.00000000 -0.07044632 0.16523250 0.16523250
#> 12: 0.00000000 -0.08418721 0.19702666 0.19702666
#> 13: 0.00000000 -0.01583434 0.03212551 0.03212551
#> 14: 0.00000000 -0.06563720 0.12429679 0.12429679
#> 15: 0.00000000 -0.08028787 0.15051289 0.15051289
#> 16: 0.00000000 -0.03892565 0.02856692 0.02856692
#> 17: 0.00000000 -0.13355386 0.09742727 0.09742727
#> 18: 0.00000000 -0.15680088 0.11418644 0.11418644
#> 19: 0.21268856  0.21268856 2.57350940 2.57350940
#> 20: 0.31415598  0.31415598 2.38692575 2.38692575
#> 21: 0.35840663  0.35840663 2.30416217 2.30416217
#> 22: 0.32511683  0.32511683 2.15548849 2.15548849
#> 23: 0.39528481  0.39528481 2.03858823 2.03858823
#> 24: 0.42753998  0.42753998 1.98480606 1.98480606
#> 25: 0.23798475  0.23798475 1.54265349 1.54265349
#> 26: 0.31071769  0.31071769 1.49142922 1.49142922
#> 27: 0.34410983  0.34410983 1.46793427 1.46793427
#>          lower      lower2      upper     upper2

## add confidence intervals computed on the log-log scale
## and backtransformed
outCI <- confint(fit.ate,
                 meanRisk.transform = "loglog", diffRisk.transform = "atanh",
                 ratioRisk.transform = "log")
summary(outCI, type = "risk", short = TRUE)
#>                        Length Class      Mode     
#> meanRisk               10     data.table list     
#> diffRisk               15     data.table list     
#> ratioRisk              15     data.table list     
#> inference              12     data.frame list     
#> inference.allContrasts  6     -none-     character
#> inference.contrasts     3     -none-     character
#> transform               3     -none-     character

dt.ate[type == "meanRisk", newse := se/(estimate*log(estimate))]
#>          type estimator time level     estimate         se       lower
#>  1:  meanRisk  GFORMULA    1    T0  0.033897209 0.02557002  0.00000000
#>  2:  meanRisk  GFORMULA    2    T0  0.135199903 0.05576405  0.02590437
#>  3:  meanRisk  GFORMULA    3    T0  0.170306018 0.06256730  0.04767637
#>  4:  meanRisk  GFORMULA    1    T1  0.047222167 0.02794464  0.00000000
#>  5:  meanRisk  GFORMULA    2    T1  0.182592994 0.05955234  0.06587254
#>  6:  meanRisk  GFORMULA    3    T1  0.226725745 0.06574061  0.09787653
#>  7:  meanRisk  GFORMULA    1    T2  0.042042798 0.02326003  0.00000000
#>  8:  meanRisk  GFORMULA    2    T2  0.164529699 0.04632579  0.07373281
#>  9:  meanRisk  GFORMULA    3    T2  0.205418524 0.05292254  0.10169224
#> 10:  diffRisk  GFORMULA    1 T0.T1  0.013324958 0.01620192 -0.01843023
#> 11:  diffRisk  GFORMULA    2 T0.T1  0.047393091 0.06012325 -0.07044632
#> 12:  diffRisk  GFORMULA    3 T0.T1  0.056419727 0.07173955 -0.08418721
#> 13:  diffRisk  GFORMULA    1 T0.T2  0.008145589 0.01223488 -0.01583434
#> 14:  diffRisk  GFORMULA    2 T0.T2  0.029329797 0.04845344 -0.06563720
#> 15:  diffRisk  GFORMULA    3 T0.T2  0.035112506 0.05887883 -0.08028787
#> 16:  diffRisk  GFORMULA    1 T1.T2 -0.005179369 0.01721781 -0.03892565
#> 17:  diffRisk  GFORMULA    2 T1.T2 -0.018063295 0.05892484 -0.13355386
#> 18:  diffRisk  GFORMULA    3 T1.T2 -0.021307221 0.06913069 -0.15680088
#> 19: ratioRisk  GFORMULA    1 T0.T1  1.393098981 0.60226128  0.21268856
#> 20: ratioRisk  GFORMULA    2 T0.T1  1.350540868 0.52877752  0.31415598
#> 21: ratioRisk  GFORMULA    3 T0.T1  1.331284402 0.49637533  0.35840663
#> 22: ratioRisk  GFORMULA    1 T0.T2  1.240302659 0.46694013  0.32511683
#> 23: ratioRisk  GFORMULA    2 T0.T2  1.216936521 0.41921776  0.39528481
#> 24: ratioRisk  GFORMULA    3 T0.T2  1.206173022 0.39726905  0.42753998
#> 25: ratioRisk  GFORMULA    1 T1.T2  0.890319121 0.33282978  0.23798475
#> 26: ratioRisk  GFORMULA    2 T1.T2  0.901073451 0.30120746  0.31071769
#> 27: ratioRisk  GFORMULA    3 T1.T2  0.906022049 0.28669518  0.34410983
#>          type estimator time level     estimate         se       lower
#>          upper quantileBand   lowerBand  upperBand   p.value adj.p.value
#>  1: 0.08401352     2.200264  0.00000000 0.09015800        NA          NA
#>  2: 0.24449544     2.200264  0.01250425 0.25789556        NA          NA
#>  3: 0.29293567     2.200264  0.03264143 0.30797061        NA          NA
#>  4: 0.10199266     2.273745  0.00000000 0.11076116        NA          NA
#>  5: 0.29931345     2.273745  0.04718615 0.31799984        NA          NA
#>  6: 0.35557496     2.273745  0.07724837 0.37620312        NA          NA
#>  7: 0.08763162     2.314972  0.00000000 0.09588910        NA          NA
#>  8: 0.25532659     2.314972  0.05728680 0.27177260        NA          NA
#>  9: 0.30914481     2.314972  0.08290433 0.32793271        NA          NA
#> 10: 0.04508014     2.134070 -0.02125108 0.04790099 0.4108318      0.5042
#> 11: 0.16523250     2.134070 -0.08091413 0.17570032 0.4305414      0.5251
#> 12: 0.19702666     2.134070 -0.09667749 0.20951695 0.4316026      0.5258
#> 13: 0.03212551     2.089872 -0.01742375 0.03371492 0.5055595      0.5898
#> 14: 0.12429679     2.089872 -0.07193169 0.13059129 0.5449669      0.6296
#> 15: 0.15051289     2.089872 -0.08793671 0.15816172 0.5509401      0.6359
#> 16: 0.02856692     2.039855 -0.04030120 0.02994247 0.7635559      0.8107
#> 17: 0.09742727     2.039855 -0.13826142 0.10213483 0.7591874      0.8062
#> 18: 0.11418644     2.039855 -0.16232380 0.11970936 0.7579176      0.8050
#> 19: 2.57350940     1.989982  0.19460974 2.59158822 0.5139464      0.5270
#> 20: 2.38692575     1.989982  0.29828302 2.40279872 0.5073773      0.5204
#> 21: 2.30416217     1.989982  0.34350632 2.31906248 0.5045122      0.5167
#> 22: 2.15548849     1.986203  0.31286476 2.16774056 0.6068097      0.6111
#> 23: 2.03858823     1.986203  0.38428493 2.04958811 0.6048216      0.6097
#> 24: 1.98480606     1.986203  0.41711602 1.99523003 0.6037776      0.6083
#> 25: 1.54265349     1.928798  0.24835759 1.53228066 0.7417472      0.7423
#> 26: 1.49142922     1.928798  0.32010500 1.48204191 0.7425841      0.7434
#> 27: 1.46793427     1.928798  0.35304486 1.45899924 0.7430648      0.7442
#>          upper quantileBand   lowerBand  upperBand   p.value adj.p.value
#>          newse
#>  1: -0.2228858
#>  2: -0.2061250
#>  3: -0.2075417
#>  4: -0.1938390
#>  5: -0.1917959
#>  6: -0.1953866
#>  7: -0.1745771
#>  8: -0.1560207
#>  9: -0.1627800
#> 10:         NA
#> 11:         NA
#> 12:         NA
#> 13:         NA
#> 14:         NA
#> 15:         NA
#> 16:         NA
#> 17:         NA
#> 18:         NA
#> 19:         NA
#> 20:         NA
#> 21:         NA
#> 22:         NA
#> 23:         NA
#> 24:         NA
#> 25:         NA
#> 26:         NA
#> 27:         NA
#>          newse
dt.ate[type == "meanRisk", .(lower = exp(-exp(log(-log(estimate)) - 1.96 * newse)),
                        upper = exp(-exp(log(-log(estimate)) + 1.96 * newse)))]
#>          lower     upper
#> 1: 0.005308092 0.1123030
#> 2: 0.049930563 0.2629084
#> 3: 0.070036723 0.3077237
#> 4: 0.011517821 0.1239456
#> 5: 0.084038215 0.3110947
#> 6: 0.113440244 0.3635412
#> 7: 0.011538684 0.1053196
#> 8: 0.086275187 0.2646901
#> 9: 0.113322713 0.3165183