Confidence Intervals and Confidence Bands for the Predicted Absolute Risk (Cumulative Incidence Function)

Confidence intervals and confidence Bands for the predicted absolute risk (cumulative incidence function).

Usage

# S3 method for predictCSC
confint(
  object,
  parm = NULL,
  level = 0.95,
  n.sim = 10000,
  absRisk.transform = "loglog",
  seed = NA,
  ...
)

Arguments

object

A predictCSC object, i.e. output of the predictCSC function.

parm

not used. For compatibility with the generic method.

level

[numeric, 0-1] Level of confidence.

n.sim

[integer, >0] the number of simulations used to compute the quantiles for the confidence bands.

absRisk.transform

[character] the transformation used to improve coverage of the confidence intervals for the predicted absolute risk in small samples. Can be "none", "log", "loglog", "cloglog".

seed

[integer, >0] seed number set before performing simulations for the confidence bands. If not given or NA no seed is set.

...

not used.

Details

The confidence bands and confidence intervals are automatically restricted to the interval [0;1].

Author

Brice Ozenne

Examples

library(survival)
library(prodlim)
#### generate data ####
set.seed(10)
d <- sampleData(100) 

#### estimate a stratified CSC model ###
fit <- CSC(Hist(time,event)~ X1 + strata(X2) + X6, data=d)

#### compute individual specific risks
fit.pred <- predict(fit, newdata=d[1:3], times=c(3,8), cause = 1,
                    se = TRUE, iid = TRUE, band = TRUE)
fit.pred
#>    observation X1   X6 X2 times strata absRisk absRisk.se absRisk.lower
#> 1:           1  1 30.8  1     3   X2=1   0.138     0.0785        0.0311
#> 2:           2  0 66.3  1     3   X2=1   0.456     0.1247        0.2114
#> 3:           3  0 77.6  1     3   X2=1   0.799     0.1285        0.4007
#> 4:           1  1 30.8  1     8   X2=1   0.298     0.1365        0.0788
#> 5:           2  0 66.3  1     8   X2=1   0.803     0.0900        0.5503
#> 6:           3  0 77.6  1     8   X2=1   0.954     0.0659        0.4310
#>    absRisk.upper absRisk.quantileBand absRisk.lowerBand absRisk.upperBand
#> 1:         0.324                 2.15            0.0255             0.344
#> 2:         0.673                 2.21            0.1832             0.695
#> 3:         0.946                 2.15            0.3498             0.953
#> 4:         0.562                 2.15            0.0649             0.585
#> 5:         0.923                 2.21            0.5067             0.932
#> 6:         0.997                 2.15            0.3266             0.998

## check confidence intervals
newse <- fit.pred$absRisk.se/(-fit.pred$absRisk*log(fit.pred$absRisk))
cbind(lower = as.double(exp(-exp(log(-log(fit.pred$absRisk)) + 1.96 * newse))),
      upper = as.double(exp(-exp(log(-log(fit.pred$absRisk)) - 1.96 * newse)))
)
#>           lower     upper
#> [1,] 0.03106980 0.3237727
#> [2,] 0.21142317 0.6725332
#> [3,] 0.40068902 0.9463311
#> [4,] 0.07874874 0.5616676
#> [5,] 0.55031414 0.9226697
#> [6,] 0.43098190 0.9973992

#### compute confidence intervals without transformation
confint(fit.pred, absRisk.transform = "none")
#>    observation X1   X6 X2 times strata absRisk absRisk.se absRisk.lower
#> 1:           1  1 30.8  1     3   X2=1   0.138     0.0785        0.0000
#> 2:           2  0 66.3  1     3   X2=1   0.456     0.1247        0.2116
#> 3:           3  0 77.6  1     3   X2=1   0.799     0.1285        0.5469
#> 4:           1  1 30.8  1     8   X2=1   0.298     0.1365        0.0305
#> 5:           2  0 66.3  1     8   X2=1   0.803     0.0900        0.6267
#> 6:           3  0 77.6  1     8   X2=1   0.954     0.0659        0.8252
#>    absRisk.upper absRisk.quantileBand absRisk.lowerBand absRisk.upperBand
#> 1:         0.292                 2.14           0.00000             0.306
#> 2:         0.700                 2.23           0.17842             0.734
#> 3:         1.000                 2.17           0.51970             1.000
#> 4:         0.565                 2.14           0.00656             0.589
#> 5:         0.980                 2.23           0.60269             1.000
#> 6:         1.000                 2.17           0.81125             1.000
cbind(lower = as.double(fit.pred$absRisk - 1.96 * fit.pred$absRisk.se),
      upper = as.double(fit.pred$absRisk + 1.96 * fit.pred$absRisk.se)
)
#>            lower     upper
#> [1,] -0.01553676 0.2920606
#> [2,]  0.21161883 0.7004983
#> [3,]  0.54689863 1.0507549
#> [4,]  0.03045984 0.5654551
#> [5,]  0.62665840 0.9795842
#> [6,]  0.82518445 1.0833380