Confidence Intervals and Confidence Bands for the predicted Survival/Cumulative Hazard

Confidence intervals and confidence Bands for the predicted survival/cumulative Hazard.

Usage

# S3 method for predictCox
confint(
  object,
  parm = NULL,
  level = 0.95,
  n.sim = 10000,
  cumhazard.transform = "log",
  survival.transform = "loglog",
  seed = NA,
  ...
)

Arguments

object

A predictCox object, i.e. output of the predictCox function.

parm

[character] the type of predicted value for which the confidence intervals should be output. Can be "survival" or "cumhazard".

level

[numeric, 0-1] Level of confidence.

n.sim

[integer, >0] the number of simulations used to compute the quantiles for the confidence bands.

cumhazard.transform

[character] the transformation used to improve coverage of the confidence intervals for the cumlative hazard in small samples. Can be "none", "log".

survival.transform

[character] the transformation used to improve coverage of the confidence intervals for the survival in small samples. Can be "none", "log", "loglog", "cloglog".

seed

[integer, >0] seed number set before performing simulations for the confidence bands. If not given or NA no seed is set.

...

not used.

Details

The confidence bands and confidence intervals are automatically restricted to the interval of definition of the statistic, i.e. a confidence interval for the survival of [0.5;1.2] will become [0.5;1].

Author

Brice Ozenne

Examples

library(survival)

#### generate data ####
set.seed(10)
d <- sampleData(40,outcome="survival") 

#### estimate a stratified Cox model ####
fit <- coxph(Surv(time,event)~X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)

#### compute individual specific survival probabilities  
fit.pred <- predictCox(fit, newdata=d[1:3], times=c(3,8), type = "survival",
                       se = TRUE, iid = TRUE, band = TRUE)
#> Error in rowMultiply_cpp(-iIFcumhazard, scale = new.survival[iObs, ]): could not find function "rowMultiply_cpp"
fit.pred
#> Error in eval(expr, envir, enclos): object 'fit.pred' not found

## check standard error
sqrt(rowSums(fit.pred$survival.iid[,,1]^2)) ## se for individual 1
#> Error in is.data.frame(x): object 'fit.pred' not found

## check confidence interval
newse <- fit.pred$survival.se/(-fit.pred$survival*log(fit.pred$survival))
#> Error in eval(expr, envir, enclos): object 'fit.pred' not found
cbind(lower = as.double(exp(-exp(log(-log(fit.pred$survival)) + 1.96 * newse))),
      upper = as.double(exp(-exp(log(-log(fit.pred$survival)) - 1.96 * newse)))
)
#> Error in cbind(lower = as.double(exp(-exp(log(-log(fit.pred$survival)) +     1.96 * newse))), upper = as.double(exp(-exp(log(-log(fit.pred$survival)) -     1.96 * newse)))): object 'fit.pred' not found

#### compute confidence intervals without transformation
confint(fit.pred, survival.transform = "none")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'confint': object 'fit.pred' not found
cbind(lower = as.double(fit.pred$survival - 1.96 * fit.pred$survival.se),
      upper = as.double(fit.pred$survival + 1.96 * fit.pred$survival.se)
)
#> Error in cbind(lower = as.double(fit.pred$survival - 1.96 * fit.pred$survival.se),     upper = as.double(fit.pred$survival + 1.96 * fit.pred$survival.se)): object 'fit.pred' not found