Input for data splitting algorithms

Parse hyperparameters for data splitting algorithm

Usage

getSplitMethod(split.method, B, N, M, seed)

Arguments

split.method

A character string specifying the algorithm for data splitting:

"loob" leave one out bootstrap

"bootcv" bootstrap cross validation

"cv5" 5-fold cross validation

"loocv" leave one out cross validation aka N-1 fold cross validation

"632plus" Efron's .632+ bootstrap

B

Number of repetitions of bootstrap or k-fold cross-validation

N

Sample size

M

Subsample size. Default is N (no subsampling).

seed

Integer passed to set.seed. If not given or NA no seed is set.

Value

A list with the following elements:

split.methodName: the print name of the algorithm

split.method: the internal name of the algorithm

index: the index for data splitting. For bootstrap splitting this is a matrix with B columns and M rows identifying the in-bag subjects. For k-fold cross-validation this is a matrix with B columns identifying the membership to the k groups.

k: the k of k-fold cross-validation

N: the sample size

M: the subsample size

See also

Author

Thomas A. Gerds <tag@biostat.ku.dk>

Examples

# 3-fold crossvalidation
getSplitMethod("cv3",B=4,N=37)
#> $name
#> [1] "3-fold cross-validation"
#> 
#> $internal.name
#> [1] "crossval"
#> 
#> $index
#>       [,1] [,2] [,3] [,4]
#>  [1,]    2    3    2    1
#>  [2,]    3    2    1    3
#>  [3,]    1    2    1    1
#>  [4,]    3    1    2    1
#>  [5,]    2    3    2    3
#>  [6,]    1    2    2    2
#>  [7,]    2    1    3    3
#>  [8,]    1    1    2    3
#>  [9,]    2    2    2    1
#> [10,]    2    1    1    3
#> [11,]    3    1    3    2
#> [12,]    3    1    3    2
#> [13,]    2    1    2    3
#> [14,]    3    2    1    2
#> [15,]    2    2    1    1
#> [16,]    3    2    2    1
#> [17,]    1    1    2    2
#> [18,]    1    1    1    1
#> [19,]    2    2    1    2
#> [20,]    1    3    1    2
#> [21,]    2    2    2    3
#> [22,]    2    2    3    1
#> [23,]    1    3    1    2
#> [24,]    3    3    1    2
#> [25,]    3    1    2    3
#> [26,]    1    3    1    3
#> [27,]    1    2    3    1
#> [28,]    3    3    3    1
#> [29,]    3    3    3    1
#> [30,]    1    2    3    3
#> [31,]    1    1    3    2
#> [32,]    1    3    1    3
#> [33,]    2    3    3    3
#> [34,]    3    1    3    2
#> [35,]    1    3    1    1
#> [36,]    3    3    3    1
#> [37,]    2    1    2    2
#> 
#> $k
#> [1] 3
#> 
#> $B
#> [1] 4
#> 
#> $M
#> [1] 37
#> 
#> $N
#> [1] 37
#> 
#> attr(,"class")
#> [1] "split.method"

# bootstrap with replacement
getSplitMethod("loob",B=4,N=37)
#> $name
#> [1] "LeaveOneOutBoot"
#> 
#> $internal.name
#> [1] "LeaveOneOutBoot"
#> 
#> $index
#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    2    1    2
#>  [2,]    2    3    2    2
#>  [3,]    3    4    2    3
#>  [4,]    5    5    3    4
#>  [5,]    6    6    3    5
#>  [6,]    8    6    6    6
#>  [7,]    9    7    6    7
#>  [8,]   11    9    7    8
#>  [9,]   11   10    9    8
#> [10,]   11   11   10    9
#> [11,]   12   11   11   11
#> [12,]   13   12   11   11
#> [13,]   13   14   15   12
#> [14,]   13   14   16   13
#> [15,]   17   14   16   15
#> [16,]   18   15   19   17
#> [17,]   19   16   19   18
#> [18,]   20   16   21   20
#> [19,]   21   16   21   20
#> [20,]   23   17   24   21
#> [21,]   24   18   24   22
#> [22,]   24   18   26   23
#> [23,]   26   19   26   23
#> [24,]   26   20   26   24
#> [25,]   28   21   27   26
#> [26,]   30   23   28   26
#> [27,]   31   24   29   27
#> [28,]   32   24   30   28
#> [29,]   32   25   32   28
#> [30,]   33   28   34   29
#> [31,]   34   29   34   29
#> [32,]   34   31   35   31
#> [33,]   34   34   35   31
#> [34,]   35   34   36   34
#> [35,]   36   35   37   34
#> [36,]   36   35   37   36
#> [37,]   37   36   37   36
#> 
#> $k
#> NULL
#> 
#> $B
#> [1] 4
#> 
#> $M
#> [1] 37
#> 
#> $N
#> [1] 37
#> 
#> attr(,"class")
#> [1] "split.method"

# bootstrap without replacement
getSplitMethod("loob",B=4,N=37,M=20)
#> $name
#> [1] "LeaveOneOutBoot"
#> 
#> $internal.name
#> [1] "LeaveOneOutBoot"
#> 
#> $index
#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    1    1    1
#>  [2,]    4    2    2    3
#>  [3,]    5    6    3    4
#>  [4,]    8    8    4    5
#>  [5,]   10   10    8    6
#>  [6,]   11   12   10    7
#>  [7,]   12   14   12    8
#>  [8,]   13   15   13    9
#>  [9,]   14   16   15   10
#> [10,]   15   17   18   12
#> [11,]   17   20   20   13
#> [12,]   18   23   21   15
#> [13,]   21   24   22   16
#> [14,]   23   25   24   17
#> [15,]   24   27   25   22
#> [16,]   26   28   29   26
#> [17,]   28   33   30   27
#> [18,]   29   34   32   28
#> [19,]   32   35   33   29
#> [20,]   35   36   34   32
#> 
#> $k
#> NULL
#> 
#> $B
#> [1] 4
#> 
#> $M
#> [1] 20
#> 
#> $N
#> [1] 37
#> 
#> attr(,"class")
#> [1] "split.method"