Fast computation of survival probabilities, hazards and cumulative hazards from Cox regression models

Fast routine to get baseline hazards and subject specific hazards as well as survival probabilities from a survival::coxph or rms::cph object

Usage

predictCox(
  object,
  times,
  newdata = NULL,
  centered = TRUE,
  type = c("cumhazard", "survival"),
  keep.strata = TRUE,
  keep.times = TRUE,
  keep.newdata = FALSE,
  keep.infoVar = FALSE,
  se = FALSE,
  band = FALSE,
  iid = FALSE,
  confint = (se + band) > 0,
  diag = FALSE,
  average.iid = FALSE,
  store.iid = "full"
)

Arguments

object

The fitted Cox regression model object either obtained with coxph (survival package) or cph (rms package).

times

[numeric vector] Time points at which to return the estimated hazard/cumulative hazard/survival.

newdata

[data.frame or data.table] Contain the values of the predictor variables defining subject specific predictions. Should have the same structure as the data set used to fit the object.

centered

[logical] If TRUE return prediction at the mean values of the covariates fit$mean, if FALSE return a prediction for all covariates equal to zero. in the linear predictor. Will be ignored if argument newdata is used. For internal use.

type

[character vector] the type of predicted value. Choices are

"hazard" the baseline hazard function when argument newdata is not used and the hazard function when argument newdata is used.

"cumhazard" the cumulative baseline hazard function when argument newdata is not used and the cumulative hazard function when argument newdata is used.

"survival" the survival baseline hazard function when argument newdata is not used and the cumulative hazard function when argument newdata is used.

Several choices can be combined in a vector of strings that match (no matter the case) strings "hazard","cumhazard", "survival".

keep.strata

[logical] If TRUE add the (newdata) strata to the output. Only if there any.

keep.times

[logical] If TRUE add the evaluation times to the output.

keep.newdata

[logical] If TRUE add the value of the covariates used to make the prediction in the output list.

keep.infoVar

[logical] For internal use.

se

[logical] If TRUE compute and add the standard errors to the output.

band

[logical] If TRUE compute and add the quantiles for the confidence bands to the output.

iid

[logical] If TRUE compute and add the influence function to the output.

confint

[logical] If TRUE compute and add the confidence intervals/bands to the output. They are computed applying the confint function to the output.

diag

[logical] when FALSE the hazard/cumlative hazard/survival for all observations at all times is computed, otherwise it is only computed for the i-th observation at the i-th time.

average.iid

[logical] If TRUE add the average of the influence function over newdata to the output.

store.iid

[character] Implementation used to estimate the influence function and the standard error. Can be "full" or "minimal".

Details

When the argument newdata is not specified, the function computes the baseline hazard estimate. See (Ozenne et al., 2017) section "Handling of tied event times".

Otherwise the function computes survival probabilities with confidence intervals/bands. See (Ozenne et al., 2017) section "Confidence intervals and confidence bands for survival probabilities". The survival is computed using the exponential approximation (equation 3).

A detailed explanation about the meaning of the argument store.iid can be found in (Ozenne et al., 2017) Appendix B "Saving the influence functions".

The function is not compatible with time varying predictor variables.

The centered argument enables us to reproduce the results obtained with the basehaz function from the survival package but should not be modified by the user.

The iid decomposition is output using an array containing the value of the influence of each subject used to fit the object (dim 1), for each subject in newdata (dim 3), and each time (dim 2).

References

Brice Ozenne, Anne Lyngholm Sorensen, Thomas Scheike, Christian Torp-Pedersen and Thomas Alexander Gerds. riskRegression: Predicting the Risk of an Event using Cox Regression Models. The R Journal (2017) 9:2, pages 440-460.

See also

Author

Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk

Examples

library(survival)
library(data.table)

#### generate data ####
set.seed(10)
d <- sampleData(40,outcome="survival") ## training dataset
nd <- sampleData(4,outcome="survival") ## validation dataset
d$time <- round(d$time,1) ## create tied events
# table(duplicated(d$time))

#### stratified Cox model ####
fit <- coxph(Surv(time,event)~X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)

## compute the baseline cumulative hazard
fit.haz <- predictCox(fit)
cbind(survival::basehaz(fit), fit.haz$cumhazard)
#>        hazard time strata fit.haz$cumhazard
#> 1  0.02740075  1.0      0        0.02740075
#> 2  0.04639643  1.8      0        0.04639643
#> 3  0.06604836  2.1      0        0.06604836
#> 4  0.10786058  2.2      0        0.10786058
#> 5  0.13034065  2.5      0        0.13034065
#> 6  0.15857763  2.7      0        0.15857763
#> 7  0.18834334  3.0      0        0.18834334
#> 8  0.18834334  3.2      0        0.18834334
#> 9  0.22364602  3.5      0        0.22364602
#> 10 0.25999786  3.6      0        0.25999786
#> 11 0.30157505  3.7      0        0.30157505
#> 12 0.34570093  4.0      0        0.34570093
#> 13 0.39935781  4.7      0        0.39935781
#> 14 0.39935781  5.1      0        0.39935781
#> 15 0.45630007  5.5      0        0.45630007
#> 16 0.51549299  5.8      0        0.51549299
#> 17 0.58000429  5.9      0        0.58000429
#> 18 0.65531323  6.6      0        0.65531323
#> 19 0.73137572  6.9      0        0.73137572
#> 20 0.73137572  7.0      0        0.73137572
#> 21 0.73137572  7.2      0        0.73137572
#> 22 0.83727365  7.7      0        0.83727365
#> 23 1.01024745  8.4      0        1.01024745
#> 24 1.01024745  9.2      0        1.01024745
#> 25 1.01024745  9.5      0        1.01024745
#> 26 1.01024745  9.6      0        1.01024745
#> 27 1.85095070 11.0      0        1.85095070
#> 28 3.88020663 11.1      0        3.88020663
#> 29 3.88020663 11.5      0        3.88020663
#> 30 8.67043704 12.0      0        8.67043704
#> 31 0.03325983  0.6      1        0.03325983
#> 32 0.15046226  3.6      1        0.15046226
#> 33 0.15046226  4.2      1        0.15046226
#> 34 0.80091039  4.7      1        0.80091039
#> 35 2.87808974  7.5      1        2.87808974

## compute individual specific cumulative hazard and survival probabilities 
fit.pred <- predictCox(fit, newdata=nd, times=c(3,8), se = TRUE, band = TRUE)
fit.pred
#>    observation strata times cumhazard cumhazard.se cumhazard.lower
#> 1:           1      1     3    0.3728       0.1390          0.1795
#> 2:           2      0     3    0.0677       0.0382          0.0224
#> 3:           3      0     3    0.1181       0.0541          0.0482
#> 4:           4      0     3    0.2030       0.0750          0.0985
#> 5:           1      1     8        NA           NA              NA
#> 6:           2      0     8    0.3008       0.1401          0.1207
#> 7:           3      0     8    0.5251       0.1885          0.2598
#> 8:           4      0     8    0.9026       0.2504          0.5240
#>    cumhazard.upper cumhazard.quantileBand cumhazard.lowerBand
#> 1:           0.774                   1.96              0.1797
#> 2:           0.205                   2.14              0.0202
#> 3:           0.290                   2.17              0.0438
#> 4:           0.419                   2.20              0.0900
#> 5:              NA                   1.96                  NA
#> 6:           0.749                   2.14              0.1110
#> 7:           1.061                   2.17              0.2412
#> 8:           1.555                   2.20              0.4899
#>    cumhazard.upperBand survival survival.se survival.lower survival.upper
#> 1:               0.773    0.689      0.0957          0.461          0.836
#> 2:               0.227    0.935      0.0357          0.815          0.978
#> 3:               0.318    0.889      0.0480          0.749          0.953
#> 4:               0.458    0.816      0.0612          0.658          0.906
#> 5:                  NA       NA          NA             NA             NA
#> 6:               0.815    0.740      0.1037          0.473          0.886
#> 7:               1.143    0.592      0.1115          0.346          0.771
#> 8:               1.663    0.406      0.1016          0.211          0.592
#>    survival.quantileBand survival.lowerBand survival.upperBand
#> 1:                  1.99              0.457              0.837
#> 2:                  2.13              0.798              0.980
#> 3:                  2.16              0.728              0.957
#> 4:                  2.20              0.633              0.914
#> 5:                  1.99                 NA                 NA
#> 6:                  2.13              0.444              0.895
#> 7:                  2.16              0.320              0.785
#> 8:                  2.20              0.190              0.612

####  other examples ####
# one strata variable
fitS <- coxph(Surv(time,event)~strata(X1)+X2,
              data=d, ties="breslow", x = TRUE, y = TRUE)

predictCox(fitS)
#>     observation strata times cumhazard survival
#>  1:           1      0   0.6    0.0241   0.9762
#>  2:           2      0   1.0    0.0743   0.9284
#>  3:           3      0   1.8    0.1007   0.9042
#>  4:           4      0   2.1    0.1278   0.8800
#>  5:           5      0   2.2    0.1836   0.8323
#>  6:           6      0   2.5    0.2140   0.8074
#>  7:           7      0   2.7    0.2454   0.7824
#>  8:           8      0   3.0    0.2778   0.7575
#>  9:           9      0   3.2    0.2778   0.7575
#> 10:          10      0   3.5    0.3124   0.7317
#> 11:          11      0   3.6    0.3483   0.7059
#> 12:          12      0   3.7    0.3855   0.6801
#> 13:          13      0   4.0    0.4241   0.6543
#> 14:          14      0   4.2    0.4241   0.6543
#> 15:          15      0   4.7    0.5102   0.6004
#> 16:          16      0   5.1    0.5102   0.6004
#> 17:          17      0   5.5    0.5611   0.5706
#> 18:          18      0   5.8    0.6148   0.5407
#> 19:          19      0   5.9    0.6716   0.5109
#> 20:          20      0   6.6    0.7356   0.4792
#> 21:          21      0   6.9    0.8040   0.4475
#> 22:          22      0   7.0    0.8040   0.4475
#> 23:          23      0   7.2    0.8040   0.4475
#> 24:          24      0   7.5    0.8900   0.4106
#> 25:          25      0   7.7    0.9900   0.3716
#> 26:          26      0   8.4    1.1011   0.3325
#> 27:          27      0   9.2    1.1011   0.3325
#> 28:          28      0   9.5    1.1011   0.3325
#> 29:          29      0   9.6    1.1011   0.3325
#> 30:          30      0  11.0    1.3511   0.2589
#> 31:          31      0  11.1    1.6845   0.1855
#> 32:          32      0  11.5    1.6845   0.1855
#> 33:          33      0  12.0    2.6845   0.0683
#> 34:          34      1   3.6    0.6161   0.5400
#>     observation strata times cumhazard survival
predictCox(fitS, newdata=nd, times = 1)
#>    observation strata times cumhazard survival
#> 1:           1      0     1    0.1205    0.886
#> 2:           2      0     1    0.0743    0.928
#> 3:           3      0     1    0.0743    0.928
#> 4:           4      0     1    0.0743    0.928

# two strata variables
set.seed(1)
d$U=sample(letters[1:5],replace=TRUE,size=NROW(d))
d$V=sample(letters[4:10],replace=TRUE,size=NROW(d))
nd$U=sample(letters[1:5],replace=TRUE,size=NROW(nd))
nd$V=sample(letters[4:10],replace=TRUE,size=NROW(nd))
fit2S <- coxph(Surv(time,event)~X1+strata(U)+strata(V)+X2,
              data=d, ties="breslow", x = TRUE, y = TRUE)
#> Warning: Loglik converged before variable  1 ; coefficient may be infinite. 

cbind(survival::basehaz(fit2S),predictCox(fit2S,type="cumhazard")$cumhazard)
#>          hazard time strata predictCox(fit2S, type = "cumhazard")$cumhazard
#> 1  0.000000e+00  5.9   a, d                                    0.000000e+00
#> 2  0.000000e+00  9.2   a, e                                    0.000000e+00
#> 3  3.333333e-01  1.0   a, g                                    3.333333e-01
#> 4  8.333333e-01  2.5   a, g                                    8.333333e-01
#> 5  8.333333e-01 11.5   a, g                                    8.333333e-01
#> 6  3.333333e-01  0.6   a, h                                    3.333333e-01
#> 7  1.333333e+00  2.7   a, h                                    1.333333e+00
#> 8  2.500000e-01  4.0   a, i                                    2.500000e-01
#> 9  5.833333e-01  4.7   a, i                                    5.833333e-01
#> 10 5.833333e-01  9.6   a, i                                    5.833333e-01
#> 11 1.000000e+00  4.7   a, j                                    1.000000e+00
#> 12 1.000000e+00  3.7   b, d                                    1.000000e+00
#> 13 0.000000e+00  7.2   b, f                                    0.000000e+00
#> 14 1.000000e+00 12.0   b, f                                    1.000000e+00
#> 15 5.000000e-01  2.2   b, g                                    5.000000e-01
#> 16 5.000000e-01  5.1   b, g                                    5.000000e-01
#> 17 5.000000e-01  6.9   b, h                                    5.000000e-01
#> 18 1.500000e+00 11.0   b, h                                    1.500000e+00
#> 19 5.000000e-01  2.1   b, i                                    5.000000e-01
#> 20 5.000000e-01  3.2   b, i                                    5.000000e-01
#> 21 1.000000e+00  6.6   b, j                                    1.000000e+00
#> 22 4.573922e-09  3.6   c, e                                    4.573922e-09
#> 23 1.000000e+00 11.1   c, e                                    1.000000e+00
#> 24 5.000000e-01  1.8   c, i                                    5.000000e-01
#> 25 5.000000e-01  7.0   c, i                                    5.000000e-01
#> 26 0.000000e+00  2.2   d, d                                    0.000000e+00
#> 27 0.000000e+00  9.5   d, d                                    0.000000e+00
#> 28 1.000000e+00  8.4   d, f                                    1.000000e+00
#> 29 1.000000e+00  1.0   d, g                                    1.000000e+00
#> 30 2.500000e-01  5.9   d, h                                    2.500000e-01
#> 31 5.833333e-01  7.5   d, h                                    5.833333e-01
#> 32 1.583333e+00  7.7   d, h                                    1.583333e+00
#> 33 0.000000e+00  4.2   d, j                                    0.000000e+00
#> 34 1.000000e+00  5.8   e, d                                    1.000000e+00
#> 35 1.000000e+00  3.0   e, f                                    1.000000e+00
#> 36 5.000000e-01  3.5   e, i                                    5.000000e-01
#> 37 1.500000e+00  5.5   e, i                                    1.500000e+00
#> 38 3.333333e-01  2.2   e, j                                    3.333333e-01
#> 39 8.333333e-01  3.6   e, j                                    8.333333e-01
#> 40 8.333333e-01  9.2   e, j                                    8.333333e-01
predictCox(fit2S)
#>     observation strata times cumhazard survival
#>  1:           1   a, d   5.9  0.00e+00    1.000
#>  2:           2   a, e   9.2  0.00e+00    1.000
#>  3:           3   a, g   1.0  3.33e-01    0.717
#>  4:           4   a, g   2.5  8.33e-01    0.435
#>  5:           5   a, g  11.5  8.33e-01    0.435
#>  6:           6   a, h   0.6  3.33e-01    0.717
#>  7:           7   a, h   2.7  1.33e+00    0.264
#>  8:           8   a, i   4.0  2.50e-01    0.779
#>  9:           9   a, i   4.7  5.83e-01    0.558
#> 10:          10   a, i   9.6  5.83e-01    0.558
#> 11:          11   a, j   4.7  1.00e+00    0.368
#> 12:          12   b, d   3.7  1.00e+00    0.368
#> 13:          13   b, f   7.2  0.00e+00    1.000
#> 14:          14   b, f  12.0  1.00e+00    0.368
#> 15:          15   b, g   2.2  5.00e-01    0.607
#> 16:          16   b, g   5.1  5.00e-01    0.607
#> 17:          17   b, h   6.9  5.00e-01    0.607
#> 18:          18   b, h  11.0  1.50e+00    0.223
#> 19:          19   b, i   2.1  5.00e-01    0.607
#> 20:          20   b, i   3.2  5.00e-01    0.607
#> 21:          21   b, j   6.6  1.00e+00    0.368
#> 22:          22   c, e   3.6  4.57e-09    1.000
#> 23:          23   c, e  11.1  1.00e+00    0.368
#> 24:          24   c, i   1.8  5.00e-01    0.607
#> 25:          25   c, i   7.0  5.00e-01    0.607
#> 26:          26   d, d   2.2  0.00e+00    1.000
#> 27:          27   d, d   9.5  0.00e+00    1.000
#> 28:          28   d, f   8.4  1.00e+00    0.368
#> 29:          29   d, g   1.0  1.00e+00    0.368
#> 30:          30   d, h   5.9  2.50e-01    0.779
#> 31:          31   d, h   7.5  5.83e-01    0.558
#> 32:          32   d, h   7.7  1.58e+00    0.205
#> 33:          33   d, j   4.2  0.00e+00    1.000
#> 34:          34   e, d   5.8  1.00e+00    0.368
#> 35:          35   e, f   3.0  1.00e+00    0.368
#> 36:          36   e, i   3.5  5.00e-01    0.607
#> 37:          37   e, i   5.5  1.50e+00    0.223
#> 38:          38   e, j   2.2  3.33e-01    0.717
#> 39:          39   e, j   3.6  8.33e-01    0.435
#> 40:          40   e, j   9.2  8.33e-01    0.435
#>     observation strata times cumhazard survival
predictCox(fitS, newdata=nd, times = 3)
#>    observation strata times cumhazard survival
#> 1:           1      0     3     0.451    0.637
#> 2:           2      0     3     0.278    0.757
#> 3:           3      0     3     0.278    0.757
#> 4:           4      0     3     0.278    0.757

# left truncation
test2 <- list(start=c(1,2,5,2,1,7,3,4,8,8), 
              stop=c(2,3,6,7,8,9,9,9,14,17), 
              event=c(1,1,1,1,1,1,1,0,0,0), 
              x=c(1,0,0,1,0,1,1,1,0,0)) 
m.cph <- coxph(Surv(start, stop, event) ~ 1, test2, x = TRUE)
as.data.table(predictCox(m.cph))
#> Warning: The current version of predictCox was not designed to handle left censoring 
#> The function may be used on own risks 
#>    observation times cumhazard  survival
#> 1:           1     2 0.5000000 0.6065307
#> 2:           2     3 0.8333333 0.4345982
#> 3:           3     6 1.0333333 0.3558189
#> 4:           4     7 1.2833333 0.2771121
#> 5:           5     8 1.5333333 0.2158151
#> 6:           6     9 1.9833333 0.1376098
#> 7:           7    14 1.9833333 0.1376098
#> 8:           8    17 1.9833333 0.1376098

basehaz(m.cph)
#>      hazard time
#> 1 0.5000000    2
#> 2 0.8333333    3
#> 3 1.0333333    6
#> 4 1.2833333    7
#> 5 1.5333333    8
#> 6 1.9833333    9
#> 7 1.9833333   14
#> 8 1.9833333   17