Computation of survival probabilities from Cox regression models using the product limit estimator.

Same as predictCox except that the survival is estimated using the product limit estimator.

Usage

predictCoxPL(
  object,
  times,
  newdata = NULL,
  type = c("cumhazard", "survival"),
  keep.strata = TRUE,
  keep.infoVar = FALSE,
  ...
)

Arguments

object

The fitted Cox regression model object either obtained with coxph (survival package) or cph (rms package).

times

[numeric vector] Time points at which to return the estimated hazard/cumulative hazard/survival.

newdata

[data.frame or data.table] Contain the values of the predictor variables defining subject specific predictions. Should have the same structure as the data set used to fit the object.

type

[character vector] the type of predicted value. Choices are

"hazard" the baseline hazard function when argument newdata is not used and the hazard function when argument newdata is used.

"cumhazard" the cumulative baseline hazard function when argument newdata is not used and the cumulative hazard function when argument newdata is used.

"survival" the survival baseline hazard function when argument newdata is not used and the cumulative hazard function when argument newdata is used.

Several choices can be combined in a vector of strings that match (no matter the case) strings "hazard","cumhazard", "survival".

keep.strata

[logical] If TRUE add the (newdata) strata to the output. Only if there any.

keep.infoVar

[logical] For internal use.

...

additional arguments to be passed to predictCox.

Details

Note: the iid and standard errors are computed using the exponential approximation.

Examples

library(survival)

#### generate data ####
set.seed(10)
d <- sampleData(40,outcome="survival")
nd <- sampleData(4,outcome="survival")
d$time <- round(d$time,1)

#### Cox model ####
fit <- coxph(Surv(time,event)~ X1 + X2 + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)

## exponential approximation
predictCox(fit, newdata = d, times = 1:5)
#>      observation times cumhazard survival
#>   1:           1     1   0.02767 9.73e-01
#>   2:           2     1   0.01904 9.81e-01
#>   3:           3     1   0.00346 9.97e-01
#>   4:           4     1   0.09274 9.11e-01
#>   5:           5     1   0.01844 9.82e-01
#>  ---                                     
#> 196:          36     5   3.99435 1.84e-02
#> 197:          37     5   2.10055 1.22e-01
#> 198:          38     5   0.04125 9.60e-01
#> 199:          39     5  29.60459 1.39e-13
#> 200:          40     5   0.68889 5.02e-01

## product limit
predictCoxPL(fit, newdata = d, times = 1:5)
#> Warning: Estimated survival outside the range [0,1]. 
#> Consider using predictCox instead of predictCoxPL. 
#>      observation times cumhazard survival
#>   1:           1     1   0.02767  0.97246
#>   2:           2     1   0.01904  0.98102
#>   3:           3     1   0.00346  0.99655
#>   4:           4     1   0.09274  0.90869
#>   5:           5     1   0.01844  0.98161
#>  ---                                     
#> 196:          36     5   3.99435  0.00262
#> 197:          37     5   2.10055  0.08913
#> 198:          38     5   0.04125  0.95949
#> 199:          39     5  29.60459  0.14730
#> 200:          40     5   0.68889  0.48818

#### stratified Cox model ####
fitS <- coxph(Surv(time,event)~ X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)

## exponential approximation
predictCox(fitS, newdata = d, times = 1:5)
#>      observation strata times cumhazard survival
#>   1:           1      0     1   0.02680 9.74e-01
#>   2:           2      0     1   0.01885 9.81e-01
#>   3:           3      0     1   0.00378 9.96e-01
#>   4:           4      0     1   0.08365 9.20e-01
#>   5:           5      0     1   0.01830 9.82e-01
#>  ---                                            
#> 196:          36      0     5   3.62189 2.67e-02
#> 197:          37      0     5   1.97830 1.38e-01
#> 198:          38      0     5   0.04899 9.52e-01
#> 199:          39      1     5  17.24685 3.23e-08
#> 200:          40      0     5   0.69291 5.00e-01

## product limit
predictCoxPL(fitS, newdata = d, times = 1:5)
#> Warning: Estimated survival outside the range [0,1]. 
#> Consider using predictCox instead of predictCoxPL. 
#>      observation strata times cumhazard survival
#>   1:           1      0     1   0.02680    0.973
#>   2:           2      0     1   0.01885    0.981
#>   3:           3      0     1   0.00378    0.996
#>   4:           4      0     1   0.08365    0.916
#>   5:           5      0     1   0.01830    0.982
#>  ---                                            
#> 196:          36      0     5   3.62189    0.012
#> 197:          37      0     5   1.97830    0.113
#> 198:          38      0     5   0.04899    0.952
#> 199:          39      1     5  17.24685    5.625
#> 200:          40      0     5   0.69291    0.489

#### fully stratified Cox model ####
fitS <- coxph(Surv(time,event)~ 1,
             data=d, ties="breslow", x = TRUE, y = TRUE)

## product limit
GS <- survfit(Surv(time,event)~1, data = d)
range(predictCoxPL(fitS)$survival - GS$surv)
#> [1] 0.000000e+00 1.110223e-16

fitS <- coxph(Surv(time,event)~ strata(X2),
             data=d, ties="breslow", x = TRUE, y = TRUE)

## product limit
GS <- survfit(Surv(time,event)~X2, data = d)
range(predictCoxPL(fitS)$survival - GS$surv)
#> [1] 0.000000e+00 2.220446e-16