Simulate data with binary or time-to-event outcome

Simulate data with binary outcome and 10 covariates.

Usage

sampleData(n,outcome="competing.risks",
formula= ~ f(X1,2)+f(X2,-0.033)+f(X3,0.4)+f(X6,.1)+f(X7,-.1)+f(X8,.5)+f(X9,-1),
          intercept=0)
sampleDataTD(n,n.intervals=5,outcome="competing.risks",
formula= ~ f(X1,2)+f(X2,-0.033)+f(X3,0.4)+f(X6,.1)+f(X7,-.1)+f(X8,.5)+f(X9,-1))

Arguments

n

Sample size

outcome

Character vector. Response variables are generated according to keywords: "binary" = binary response, "survival" = survival response, "competing.risks" = competing risks response

formula

Specify regression coefficients

intercept

For binary outcome the intercept of the logistic regression.

n.intervals

sampleDataTD only: the maximum number of episodes in which the covariates are updated.

Value

Simulated data as data.table with n rows and the following columns: Y (binary outcome), time (non-binary outcome), event (non-binary outcome), X1-X5 (binary predictors), X6-X10 (continous predictors)

Details

For the actual lava::regression parameters see the function definition.

See also

Author

Thomas A. Gerds <tag@biostat.ku.dk>

Examples

sampleData(10,outcome="binary")
#>           X6       X7          X8         X9        X10 X1 X2 X3 X4 X5 Y
#>  1: 46.82367 67.24533  0.76601117  1.3039815 -1.5953831  0  0  1  0  0 0
#>  2: 77.69055 61.02286 -1.25070514 -0.2387592 -0.3046539  1  0  0  1  0 1
#>  3: 47.43607 68.04548 -1.93934626  0.6812340  0.4854172  0  0  1  0  1 0
#>  4: 45.74745 69.19576 -2.08930856  1.5028794 -1.2190607  0  0  1  1  0 0
#>  5: 47.23630 67.74454  0.41779689  0.4520591  1.3879925  0  1  1  1  1 0
#>  6: 83.33275 63.70914  0.44404509 -0.1358663 -1.4756768  0  0  1  0  0 1
#>  7: 39.19116 63.04737 -0.14412019  0.7608121 -0.2618844  0  1  0  0  1 0
#>  8: 67.81334 64.13697  0.09318769 -0.7907372  0.1239023  0  1  0  1  1 1
#>  9: 66.32582 57.58020  0.71205220  1.5738184 -1.3145448  0  1  0  0  1 0
#> 10: 65.49821 58.47139 -1.44213677  1.8802128 -0.0988130  0  0  1  1  1 0
sampleData(10,outcome="survival")
#>            X6       X7         X8         X9        X10 X1 X2 X3 X4 X5
#>  1: 105.34595 58.19687 -2.3673196  1.1758264 -0.2662620  0  0  0  1  1
#>  2:  90.71622 50.04183  0.7468906  0.1840539  0.9415957  0  1  0  0  1
#>  3:  59.69159 60.70045 -1.3214394  0.8982531  0.6179547  0  0  0  0  1
#>  4:  26.55609 55.80982 -0.1839103 -0.4361993  0.2406645  0  0  0  0  0
#>  5:  44.86785 64.13566  0.3702102  0.9412698 -0.2811275  0  0  1  1  1
#>  6:  70.93943 59.31365  0.4427237 -1.4587915 -0.1273632  0  0  1  1  0
#>  7:  44.41888 60.26168  0.3243719  1.6640414  1.4109850  0  1  0  0  0
#>  8:  69.23317 59.13929 -0.2946856 -0.1398363  2.1002374  0  1  0  0  0
#>  9:  78.25618 62.90142 -1.0414508  0.7459810 -0.1347690  0  0  0  0  1
#> 10:  62.46170 67.39584 -1.8278110  0.2303515 -1.7317539  0  0  1  0  0
#>      eventtime  censtime       time event
#>  1:  2.1584938 12.291536  2.1584938     1
#>  2:  0.7952726  7.687468  0.7952726     1
#>  3:  9.6945964 13.932777  9.6945964     1
#>  4: 48.0137416  7.058650  7.0586503     0
#>  5: 55.3383918 11.390308 11.3903079     0
#>  6:  2.2579616 15.292583  2.2579616     1
#>  7: 23.5150791  9.408254  9.4082539     0
#>  8:  2.3798579  7.207902  2.3798579     1
#>  9: 12.2951520  2.611471  2.6114709     0
#> 10: 12.1308164 16.270998 12.1308164     1
sampleData(10,outcome="competing.risks")
#>           X6       X7          X8          X9        X10 X1 X2 X3 X4 X5
#>  1: 70.95693 69.90523 -0.11716037 -1.60872843 -0.8276721  0  0  0  0  1
#>  2: 48.44728 60.88304 -0.92216042  0.75254134 -0.6963852  0  0  0  1  1
#>  3: 45.47900 60.71565  0.30189182 -0.93065381  1.9186075  0  1  0  0  1
#>  4: 64.37623 63.43177  1.35253949 -0.41297456 -0.7568721  1  0  0  0  1
#>  5: 76.89589 54.65964  0.93818152 -0.58867640  1.4836873  0  1  0  1  0
#>  6: 36.21012 58.08425  0.27056390 -0.72435849 -0.1764810  0  1  0  1  0
#>  7: 55.36911 63.40905  1.27362824 -0.35314014  1.5612745  0  0  0  0  1
#>  8: 52.17097 65.28943 -1.16935075 -0.72582011 -1.4857048  0  0  0  0  0
#>  9: 68.56586 61.42335  0.08619538 -0.02632572 -0.5342470  0  0  0  0  0
#> 10: 59.30486 57.85724 -0.58657058 -0.15254579 -0.2432629  0  0  1  0  0
#>     eventtime1 eventtime2  censtime       time event
#>  1:  5.0641127  15.504000  6.941363  5.0641127     1
#>  2: 61.5754692  12.371662 13.492984 12.3716622     2
#>  3: 18.1930468  12.201308 14.666782 12.2013077     2
#>  4:  0.8187567   6.553765  9.042609  0.8187567     1
#>  5:  1.8570417   9.563060  4.970838  1.8570417     1
#>  6: 14.1247074   4.960097 10.476599  4.9600967     2
#>  7:  6.8519327   9.906128  8.926625  6.8519327     1
#>  8:  4.2771898   6.811277  8.223740  4.2771898     1
#>  9:  3.6496208   5.146281  3.941413  3.6496208     1
#> 10:  9.0028432   4.160078  9.114821  4.1600785     2