Logistic Regression Using IPCW

Logistic regression over multiple timepoints where right-censoring is handle using inverse probability of censoring weighting.

Usage

wglm(
  regressor.event,
  formula.censor,
  times,
  data,
  cause = NA,
  fitter = "coxph",
  product.limit = FALSE
)

Arguments

regressor.event

[formula] a formula with empty left hand side and the covariates for the logistic regression on the right hand side.

formula.censor

[formula] a formula used to fit the censoring model.

times

[numeric vector] time points at which to model the probability of experiencing an event.

data

[data.frame] dataset containing the time at which the event occured, the type of event, and regressors used to fit the censoring and logistic models.

cause

[character or numeric] the cause of interest. Defaults to the first cause.

fitter

[character] routine to fit the Cox regression models.

product.limit

[logical] if TRUE the survival is computed using the product limit estimator.

Value

an object of class "wglm".

Details

First, a Cox model is fitted (argument formula.censor) and the censoring probabilities are computed relative to each timepoint (argument times) to obtain the censoring weights. Then, for each timepoint, a logistic regression is fitted with the appropriate censoring weights and where the outcome is the indicator of having experience the event of interest (argument cause) at or before the timepoint.

Examples

library(survival)

set.seed(10)
n <- 250
tau <- 1:5
d <- sampleData(n, outcome = "competing.risks")
dFull <- d[event!=0] ## remove censoring
dSurv <- d[event!=2] ## remove competing risk

#### no censoring ####
e.wglm <- wglm(regressor.event = ~ X1, formula.censor = Surv(time,event==0) ~ 1,
               times = tau, data = dFull, product.limit = TRUE)
e.wglm ## same as a logistic regression
#>      logistic regression for cause 1: 
#> ----------------------------------------------------------------------------------
#>   - time: 1
#> 
#> Call:  glm(XX_event.1_XX ~ X1, family = binomial(link = "logit"))
#> 
#> Coefficients:
#> (Intercept)          X11  
#>      -2.430        1.226  
#> 
#> Degrees of Freedom: 185 Total (i.e. Null);  184 Residual
#> Null Deviance:	    113.7 
#> Residual Deviance: 111.3 	AIC: 115.3
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 2
#> 
#> Call:  glm(XX_event.2_XX ~ X1, family = binomial(link = "logit"))
#> 
#> Coefficients:
#> (Intercept)          X11  
#>      -1.372        0.561  
#> 
#> Degrees of Freedom: 185 Total (i.e. Null);  184 Residual
#> Null Deviance:	    191 
#> Residual Deviance: 190.3 	AIC: 194.3
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 3
#> 
#> Call:  glm(XX_event.3_XX ~ X1, family = binomial(link = "logit"))
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -0.8445       0.6904  
#> 
#> Degrees of Freedom: 185 Total (i.e. Null);  184 Residual
#> Null Deviance:	    230.8 
#> Residual Deviance: 229.5 	AIC: 233.5
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 4
#> 
#> Call:  glm(XX_event.4_XX ~ X1, family = binomial(link = "logit"))
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -0.4832       0.6373  
#> 
#> Degrees of Freedom: 185 Total (i.e. Null);  184 Residual
#> Null Deviance:	    249.2 
#> Residual Deviance: 248 	AIC: 252
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 5
#> 
#> Call:  glm(XX_event.5_XX ~ X1, family = binomial(link = "logit"))
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -0.1506       0.9615  
#> 
#> Degrees of Freedom: 185 Total (i.e. Null);  184 Residual
#> Null Deviance:	    257.5 
#> Residual Deviance: 254.9 	AIC: 258.9
#> ----------------------------------------------------------------------------------

summary(ate(e.wglm, data = dFull, times = tau, treatment = "X1", verbose = FALSE))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'summary': could not find function "colMultiply_cpp"

#### right-censoring ####
## no covariante in the censoring model (independent censoring)
eC.wglm <- wglm(regressor.event = ~ X1, formula.censor = Surv(time,event==0) ~ 1,
               times = tau, data = dSurv, product.limit = TRUE)
eC.wglm
#>      IPCW logistic regression : 
#> ----------------------------------------------------------------------------------
#>   - time: 1
#> 
#> Call:  glm(XX_event.1_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.1_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>       -2.36         1.26  
#> 
#> Degrees of Freedom: 173 Total (i.e. Null);  172 Residual
#> Null Deviance:	    111.9 
#> Residual Deviance: 109.3 	AIC: 112.8
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 2
#> 
#> Call:  glm(XX_event.2_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.2_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>      -1.277        0.560  
#> 
#> Degrees of Freedom: 169 Total (i.e. Null);  168 Residual
#> Null Deviance:	    186.9 
#> Residual Deviance: 186.2 	AIC: 186.4
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 3
#> 
#> Call:  glm(XX_event.3_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.3_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -0.7090       0.6523  
#> 
#> Degrees of Freedom: 163 Total (i.e. Null);  162 Residual
#> Null Deviance:	    224.6 
#> Residual Deviance: 223.4 	AIC: 216
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 4
#> 
#> Call:  glm(XX_event.4_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.4_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -0.2975       0.6894  
#> 
#> Degrees of Freedom: 152 Total (i.e. Null);  151 Residual
#> Null Deviance:	    239.9 
#> Residual Deviance: 238.6 	AIC: 215.5
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 5
#> 
#> Call:  glm(XX_event.5_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.5_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>      0.1487       1.8171  
#> 
#> Degrees of Freedom: 143 Total (i.e. Null);  142 Residual
#> Null Deviance:	    240.1 
#> Residual Deviance: 234.4 	AIC: 193.2
#> ----------------------------------------------------------------------------------

## with covariates in the censoring model
eC2.wglm <- wglm(regressor.event = ~ X1 + X8,
                formula.censor = Surv(time,event==0) ~ X1*X8,
                 times = tau, data = dSurv)
eC2.wglm
#>      IPCW logistic regression : 
#> ----------------------------------------------------------------------------------
#>   - time: 1
#> 
#> Call:  glm(XX_event.1_XX ~ X1 + X8, family = binomial(link = "logit"), 
#>     weights = "XX_IPCW.1_XX")
#> 
#> Coefficients:
#> (Intercept)          X11           X8  
#>      -3.101        1.243        1.077  
#> 
#> Degrees of Freedom: 173 Total (i.e. Null);  171 Residual
#> Null Deviance:	    111.8 
#> Residual Deviance: 92.42 	AIC: 98.07
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 2
#> 
#> Call:  glm(XX_event.2_XX ~ X1 + X8, family = binomial(link = "logit"), 
#>     weights = "XX_IPCW.2_XX")
#> 
#> Coefficients:
#> (Intercept)          X11           X8  
#>     -1.5679       0.4646       0.6977  
#> 
#> Degrees of Freedom: 169 Total (i.e. Null);  167 Residual
#> Null Deviance:	    186.8 
#> Residual Deviance: 170.3 	AIC: 173.1
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 3
#> 
#> Call:  glm(XX_event.3_XX ~ X1 + X8, family = binomial(link = "logit"), 
#>     weights = "XX_IPCW.3_XX")
#> 
#> Coefficients:
#> (Intercept)          X11           X8  
#>     -0.9496       0.5895       0.7033  
#> 
#> Degrees of Freedom: 163 Total (i.e. Null);  161 Residual
#> Null Deviance:	    224.3 
#> Residual Deviance: 203.1 	AIC: 199.4
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 4
#> 
#> Call:  glm(XX_event.4_XX ~ X1 + X8, family = binomial(link = "logit"), 
#>     weights = "XX_IPCW.4_XX")
#> 
#> Coefficients:
#> (Intercept)          X11           X8  
#>     -0.4363       0.6414       0.5335  
#> 
#> Degrees of Freedom: 152 Total (i.e. Null);  150 Residual
#> Null Deviance:	    239.8 
#> Residual Deviance: 224.7 	AIC: 206.8
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 5
#> 
#> Call:  glm(XX_event.5_XX ~ X1 + X8, family = binomial(link = "logit"), 
#>     weights = "XX_IPCW.5_XX")
#> 
#> Coefficients:
#> (Intercept)          X11           X8  
#>     0.06096      1.67334      0.45678  
#> 
#> Degrees of Freedom: 143 Total (i.e. Null);  141 Residual
#> Null Deviance:	    239.7 
#> Residual Deviance: 224 	AIC: 199.9
#> ----------------------------------------------------------------------------------

#### Competing risks ####
## here Kaplan-Meier as censoring model
eCR.wglm <- wglm(regressor.event = ~ X1,
                 formula.censor = Surv(time,event==0) ~ strata(X1),
                 times = tau, data = d, cause = 1, product.limit = TRUE)
eCR.wglm
#>      IPCW logistic regression for cause 1: 
#> ----------------------------------------------------------------------------------
#>   - time: 1
#> 
#> Call:  glm(XX_event.1_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.1_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>      -2.756        1.370  
#> 
#> Degrees of Freedom: 248 Total (i.e. Null);  247 Residual
#> Null Deviance:	    124.4 
#> Residual Deviance: 121.4 	AIC: 125
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 2
#> 
#> Call:  glm(XX_event.2_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.2_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>      -1.732        0.720  
#> 
#> Degrees of Freedom: 244 Total (i.e. Null);  243 Residual
#> Null Deviance:	    217.6 
#> Residual Deviance: 216.4 	AIC: 217.6
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 3
#> 
#> Call:  glm(XX_event.3_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.3_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -1.2312       0.8257  
#> 
#> Degrees of Freedom: 238 Total (i.e. Null);  237 Residual
#> Null Deviance:	    273.5 
#> Residual Deviance: 271.3 	AIC: 267
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 4
#> 
#> Call:  glm(XX_event.4_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.4_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -0.8904       0.7569  
#> 
#> Degrees of Freedom: 227 Total (i.e. Null);  226 Residual
#> Null Deviance:	    306.1 
#> Residual Deviance: 304.2 	AIC: 288.2
#> ----------------------------------------------------------------------------------
#> ----------------------------------------------------------------------------------
#>   - time: 5
#> 
#> Call:  glm(XX_event.5_XX ~ X1, family = binomial(link = "logit"), weights = "XX_IPCW.5_XX")
#> 
#> Coefficients:
#> (Intercept)          X11  
#>     -0.5634       1.1743  
#> 
#> Degrees of Freedom: 218 Total (i.e. Null);  217 Residual
#> Null Deviance:	    332 
#> Residual Deviance: 327.3 	AIC: 301
#> ----------------------------------------------------------------------------------