Confidence Intervals and Confidence Bands for the predicted Survival/Cumulative Hazard

Confidence intervals and confidence Bands for the predicted survival/cumulative Hazard.

Usage

# S3 method for predictCox
confint(
  object,
  parm = NULL,
  level = 0.95,
  n.sim = 10000,
  cumhazard.transform = "log",
  survival.transform = "loglog",
  seed = NA,
  ...
)

Arguments

object

A predictCox object, i.e. output of the predictCox function.

parm

[character] the type of predicted value for which the confidence intervals should be output. Can be "survival" or "cumhazard".

level

[numeric, 0-1] Level of confidence.

n.sim

[integer, >0] the number of simulations used to compute the quantiles for the confidence bands.

cumhazard.transform

[character] the transformation used to improve coverage of the confidence intervals for the cumlative hazard in small samples. Can be "none", "log".

survival.transform

[character] the transformation used to improve coverage of the confidence intervals for the survival in small samples. Can be "none", "log", "loglog", "cloglog".

seed

[integer, >0] seed number set before performing simulations for the confidence bands. If not given or NA no seed is set.

...

not used.

Details

The confidence bands and confidence intervals are automatically restricted to the interval of definition of the statistic, i.e. a confidence interval for the survival of [0.5;1.2] will become [0.5;1].

Author

Brice Ozenne

Examples

library(survival)

#### generate data ####
set.seed(10)
d <- sampleData(40,outcome="survival") 

#### estimate a stratified Cox model ####
fit <- coxph(Surv(time,event)~X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)

#### compute individual specific survival probabilities  
fit.pred <- predictCox(fit, newdata=d[1:3], times=c(3,8), type = "survival",
                       se = TRUE, iid = TRUE, band = TRUE)
fit.pred
#>    observation strata times survival survival.se survival.lower survival.upper
#> 1:           1      0     3    0.830      0.0598          0.672          0.917
#> 2:           2      0     3    0.878      0.0510          0.732          0.947
#> 3:           3      0     3    0.975      0.0194          0.888          0.995
#> 4:           1      0     8    0.441      0.1062          0.233          0.631
#> 5:           2      0     8    0.564      0.1131          0.320          0.750
#> 6:           3      0     8    0.894      0.0692          0.648          0.972
#>    survival.quantileBand survival.lowerBand survival.upperBand
#> 1:                  2.24              0.642              0.925
#> 2:                  2.22              0.704              0.953
#> 3:                  2.13              0.873              0.995
#> 4:                  2.24              0.205              0.654
#> 5:                  2.22              0.287              0.769
#> 6:                  2.13              0.613              0.975

## check standard error
sqrt(rowSums(fit.pred$survival.iid[,,1]^2)) ## se for individual 1
#>  [1] 0.018392578 0.016237776 0.007431816 0.012893439 0.025032667 0.005398069
#>  [7] 0.025666528 0.005026215 0.020439989 0.013954960 0.020245964 0.012961149
#> [13] 0.010320665 0.022387257 0.003809652 0.037842644 0.013409129 0.010237400
#> [19] 0.017079206 0.014556417 0.004868714 0.013372745 0.051035469 0.019664220
#> [25] 0.023187719 0.013702212 0.001694092 0.020826096 0.002277766 0.060597775
#> [31] 0.015962232 0.011798179 0.008106319 0.004469016 0.004496629 0.002184175
#> [37] 0.007678302 0.000496456 0.002440726 0.008088248

## check confidence interval
newse <- fit.pred$survival.se/(-fit.pred$survival*log(fit.pred$survival))
cbind(lower = as.double(exp(-exp(log(-log(fit.pred$survival)) + 1.96 * newse))),
      upper = as.double(exp(-exp(log(-log(fit.pred$survival)) - 1.96 * newse)))
)
#>          lower     upper
#> [1,] 0.6720145 0.9166403
#> [2,] 0.7317857 0.9473215
#> [3,] 0.8882076 0.9945995
#> [4,] 0.2325654 0.6310395
#> [5,] 0.3204751 0.7495790
#> [6,] 0.6477553 0.9717133

#### compute confidence intervals without transformation
confint(fit.pred, survival.transform = "none")
#>    observation strata times survival survival.se survival.lower survival.upper
#> 1:           1      0     3    0.830      0.0598          0.713          0.948
#> 2:           2      0     3    0.878      0.0510          0.778          0.978
#> 3:           3      0     3    0.975      0.0194          0.937          1.000
#> 4:           1      0     8    0.441      0.1062          0.232          0.649
#> 5:           2      0     8    0.564      0.1131          0.342          0.786
#> 6:           3      0     8    0.894      0.0692          0.759          1.000
#>    survival.quantileBand survival.lowerBand survival.upperBand
#> 1:                  2.24              0.697              0.964
#> 2:                  2.20              0.766              0.991
#> 3:                  2.10              0.934              1.000
#> 4:                  2.24              0.203              0.678
#> 5:                  2.20              0.315              0.813
#> 6:                  2.10              0.749              1.000
cbind(lower = as.double(fit.pred$survival - 1.96 * fit.pred$survival.se),
      upper = as.double(fit.pred$survival + 1.96 * fit.pred$survival.se)
)
#>          lower     upper
#> [1,] 0.7130004 0.9475453
#> [2,] 0.7780635 0.9781347
#> [3,] 0.9368629 1.0130999
#> [4,] 0.2324617 0.6488784
#> [5,] 0.3422163 0.7857635
#> [6,] 0.7587567 1.0300007