Input for data splitting algorithms

Parse hyperparameters for data splitting algorithm

Usage

getSplitMethod(split.method, B, N, M, seed)

Arguments

split.method

A character string specifying the algorithm for data splitting:

"loob" leave one out bootstrap

"bootcv" bootstrap cross validation

"cv5" 5-fold cross validation

"loocv" leave one out cross validation aka N-1 fold cross validation

"632plus" Efron's .632+ bootstrap

B

Number of repetitions of bootstrap or k-fold cross-validation

N

Sample size

M

Subsample size. Default is N (no subsampling).

seed

Integer passed to set.seed. If not given or NA no seed is set.

Value

A list with the following elements:

split.methodName: the print name of the algorithm

split.method: the internal name of the algorithm

index: the index for data splitting. For bootstrap splitting this is a matrix with B columns and M rows identifying the in-bag subjects. For k-fold cross-validation this is a matrix with B columns identifying the membership to the k groups.

k: the k of k-fold cross-validation

N: the sample size

M: the subsample size

See also

Author

Thomas A. Gerds <tag@biostat.ku.dk>

Examples

# 3-fold crossvalidation
getSplitMethod("cv3",B=4,N=37)
#> $name
#> [1] "3-fold cross-validation"
#> 
#> $internal.name
#> [1] "crossval"
#> 
#> $index
#>       [,1] [,2] [,3] [,4]
#>  [1,]    3    2    3    1
#>  [2,]    3    3    1    1
#>  [3,]    1    3    2    1
#>  [4,]    3    1    2    1
#>  [5,]    2    1    1    2
#>  [6,]    3    2    2    1
#>  [7,]    1    2    3    1
#>  [8,]    2    1    3    3
#>  [9,]    2    3    1    1
#> [10,]    1    1    2    1
#> [11,]    2    2    3    2
#> [12,]    2    1    2    3
#> [13,]    3    2    2    1
#> [14,]    3    1    1    2
#> [15,]    3    1    1    3
#> [16,]    1    1    3    3
#> [17,]    3    1    2    3
#> [18,]    3    3    2    3
#> [19,]    1    3    1    2
#> [20,]    2    1    2    2
#> [21,]    1    1    1    3
#> [22,]    1    3    3    2
#> [23,]    2    2    1    3
#> [24,]    1    3    1    3
#> [25,]    1    3    1    2
#> [26,]    2    2    2    1
#> [27,]    2    1    1    2
#> [28,]    3    3    1    1
#> [29,]    1    2    2    2
#> [30,]    1    2    3    1
#> [31,]    3    1    3    3
#> [32,]    1    2    1    3
#> [33,]    2    2    3    3
#> [34,]    2    2    2    2
#> [35,]    2    3    3    2
#> [36,]    1    3    3    1
#> [37,]    3    3    3    2
#> 
#> $k
#> [1] 3
#> 
#> $B
#> [1] 4
#> 
#> $M
#> [1] 37
#> 
#> $N
#> [1] 37
#> 
#> attr(,"class")
#> [1] "split.method"

# bootstrap with replacement
getSplitMethod("loob",B=4,N=37)
#> $name
#> [1] "LeaveOneOutBoot"
#> 
#> $internal.name
#> [1] "LeaveOneOutBoot"
#> 
#> $index
#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    1    1    1
#>  [2,]    2    3    1    2
#>  [3,]    3    4    3    3
#>  [4,]    4    4    4    4
#>  [5,]    5    5    4    4
#>  [6,]    5    5    4    6
#>  [7,]    5    6    5    6
#>  [8,]    5    7    6    7
#>  [9,]    9    7    7    8
#> [10,]   10    9    8    9
#> [11,]   10   10    8   10
#> [12,]   11   10    8   10
#> [13,]   11   11    9   10
#> [14,]   15   12   12   11
#> [15,]   15   12   12   12
#> [16,]   15   13   14   12
#> [17,]   15   14   16   14
#> [18,]   16   14   16   15
#> [19,]   16   14   17   16
#> [20,]   17   14   19   17
#> [21,]   19   17   20   18
#> [22,]   21   17   20   18
#> [23,]   23   17   24   19
#> [24,]   23   18   24   19
#> [25,]   23   19   25   19
#> [26,]   26   25   25   20
#> [27,]   26   26   26   22
#> [28,]   30   27   28   22
#> [29,]   31   27   29   22
#> [30,]   34   29   30   23
#> [31,]   34   29   30   23
#> [32,]   35   31   31   26
#> [33,]   36   32   31   26
#> [34,]   37   33   31   27
#> [35,]   37   34   32   31
#> [36,]   37   36   33   36
#> [37,]   37   36   36   37
#> 
#> $k
#> NULL
#> 
#> $B
#> [1] 4
#> 
#> $M
#> [1] 37
#> 
#> $N
#> [1] 37
#> 
#> attr(,"class")
#> [1] "split.method"

# bootstrap without replacement
getSplitMethod("loob",B=4,N=37,M=20)
#> $name
#> [1] "LeaveOneOutBoot"
#> 
#> $internal.name
#> [1] "LeaveOneOutBoot"
#> 
#> $index
#>       [,1] [,2] [,3] [,4]
#>  [1,]    2    1    2    2
#>  [2,]    4    3    4    3
#>  [3,]    5    4    5    4
#>  [4,]    6    5    6    8
#>  [5,]    7   10    8    9
#>  [6,]    9   12   10   10
#>  [7,]   11   13   14   13
#>  [8,]   12   14   16   14
#>  [9,]   14   15   18   16
#> [10,]   15   17   19   19
#> [11,]   16   19   20   20
#> [12,]   18   22   22   21
#> [13,]   19   23   24   22
#> [14,]   24   25   25   23
#> [15,]   25   29   26   24
#> [16,]   26   30   29   25
#> [17,]   28   31   31   30
#> [18,]   30   34   33   31
#> [19,]   34   35   34   33
#> [20,]   35   37   35   34
#> 
#> $k
#> NULL
#> 
#> $B
#> [1] 4
#> 
#> $M
#> [1] 20
#> 
#> $N
#> [1] 37
#> 
#> attr(,"class")
#> [1] "split.method"