Simulate data with binary or time-to-event outcome

Simulate data with binary outcome and 10 covariates.

Usage

sampleData(n,outcome="competing.risks",
formula= ~ f(X1,2)+f(X2,-0.033)+f(X3,0.4)+f(X6,.1)+f(X7,-.1)+f(X8,.5)+f(X9,-1),
          intercept=0)
sampleDataTD(n,n.intervals=5,outcome="competing.risks",
formula= ~ f(X1,2)+f(X2,-0.033)+f(X3,0.4)+f(X6,.1)+f(X7,-.1)+f(X8,.5)+f(X9,-1))

Arguments

n

Sample size

outcome

Character vector. Response variables are generated according to keywords: "binary" = binary response, "survival" = survival response, "competing.risks" = competing risks response

formula

Specify regression coefficients

intercept

For binary outcome the intercept of the logistic regression.

n.intervals

sampleDataTD only: the maximum number of episodes in which the covariates are updated.

Value

Simulated data as data.table with n rows and the following columns: Y (binary outcome), time (non-binary outcome), event (non-binary outcome), X1-X5 (binary predictors), X6-X10 (continous predictors)

Details

For the actual lava::regression parameters see the function definition.

See also

Author

Thomas A. Gerds <tag@biostat.ku.dk>

Examples

set.seed(10)
sampleData(10,outcome="binary")
#>           X6       X7          X8         X9         X10 X1 X2 X3 X4 X5 Y
#>  1: 78.24189 63.93171 -0.02881534 -1.4355144 -0.37366156  0  0  1  1  1 1
#>  2: 71.03536 55.48894  0.23252515  0.3620872 -0.68755543  0  0  0  0  0 1
#>  3: 52.78187 62.66448 -0.30120868 -1.7590868 -0.87215883  0  0  1  0  0 1
#>  4: 68.44117 56.77053 -0.67761458 -0.3245440 -0.10176101  0  1  0  0  1 1
#>  5: 41.30520 61.45494  0.65522764 -0.6515630 -0.25378053  0  0  0  0  0 0
#>  6: 65.71383 53.81203 -0.40063755  1.0865514 -1.85374045  0  0  1  0  0 1
#>  7: 38.54359 57.71912 -0.33455657 -0.7625449 -0.07794607  0  1  1  0  0 0
#>  8: 44.27332 55.84839  1.36795395 -0.8286625  0.96856634  0  1  0  0  0 1
#>  9: 56.72245 61.70058  2.13776710  0.8344739  0.18492596  0  0  1  0  1 0
#> 10: 37.65096 65.33188  0.50581926 -0.9676520 -1.37994358  0  0  0  1  0 0
sampleData(10,outcome="survival")
#>           X6       X7         X8          X9         X10 X1 X2 X3 X4 X5
#>  1: 54.13344 51.62334  1.4292128 -0.48136561 -1.28015460  0  0  0  0  1
#>  2: 56.25199 53.97407  0.6334359  0.20288178  1.12886823  0  0  0  0  1
#>  3: 77.32657 50.18374 -1.9968156 -0.03173974 -0.46413453  0  0  0  0  1
#>  4: 47.02909 67.35376 -0.6818322 -1.19558030 -0.31576021  0  1  0  0  1
#>  5: 46.99982 61.86236 -0.4600555  0.62368124  0.92429315  0  0  1  0  1
#>  6: 25.18474 65.32940 -0.9830692 -0.91480448  0.07714472  0  0  0  1  0
#>  7: 69.13245 62.65325  0.4953317  0.24875801  1.03992361  0  0  0  1  0
#>  8: 77.25009 60.50992  0.7258175 -1.06262279  0.74188621  0  1  0  1  0
#>  9: 42.00603 66.68891  0.6672987 -0.36398225  1.25554486  0  0  1  1  0
#> 10: 36.29999 60.43617  0.9547864 -1.20699485  0.95091897  0  1  1  0  0
#>      eventtime  censtime       time event
#>  1:  2.6426569  5.690865  2.6426569     1
#>  2:  6.1908753  8.129396  6.1908753     1
#>  3:  4.6347317  9.417940  4.6347317     1
#>  4:  0.8593716  9.400349  0.8593716     1
#>  5: 15.7908326  5.212969  5.2129689     0
#>  6: 70.0259799 11.620008 11.6200077     0
#>  7:  8.6180977  5.198544  5.1985435     0
#>  8:  2.1205189  1.332360  1.3323596     0
#>  9: 20.8842121  6.326926  6.3269260     0
#> 10:  2.2870935  7.264675  2.2870935     1
sampleData(10,outcome="competing.risks")
#>           X6       X7          X8         X9        X10 X1 X2 X3 X4 X5
#>  1: 84.07671 58.60588 -0.32922472  0.1418803  1.2155138  0  0  1  1  1
#>  2: 71.16636 53.66634 -0.28282162  1.2617151  0.3308765  0  0  0  0  1
#>  3: 72.93123 58.75426  0.43242913 -0.4315003  1.3902751  0  1  0  1  1
#>  4: 65.92734 60.08994 -0.30760710 -1.8227126  0.8720470  0  1  0  0  1
#>  5: 67.63678 61.88536 -0.05663631  0.3525440 -1.0808170  0  0  0  0  0
#>  6: 58.16175 63.98004  0.73351542 -1.3484514  0.4958216  0  0  0  1  0
#>  7: 61.38877 55.79661  0.09731162  0.7076883  1.0526276  0  0  1  0  1
#>  8: 54.63180 48.97264  1.63089174 -0.4108909 -1.2746500  0  0  0  0  1
#>  9: 54.60517 54.35972  0.56061070 -0.4460452 -0.1936667  0  0  0  1  1
#> 10: 75.42856 53.29345  1.32956476 -1.0411563 -1.2950836  1  0  0  1  1
#>     eventtime1 eventtime2  censtime      time event
#>  1:  1.0392993  14.346374 11.789377 1.0392993     1
#>  2:  0.9393134   5.160939  7.372990 0.9393134     1
#>  3:  1.5912116  10.397899  4.299037 1.5912116     1
#>  4:  3.6640553  14.591833 12.606763 3.6640553     1
#>  5: 14.7176760   7.647145 16.262451 7.6471451     2
#>  6:  4.3279139  14.089344  5.350291 4.3279139     1
#>  7: 11.2254895   3.944821  7.463531 3.9448210     2
#>  8:  6.4158478   4.095935  5.463787 4.0959352     2
#>  9:  8.1103217  11.218448 16.068604 8.1103217     1
#> 10:  0.1623599   4.589312  2.823758 0.1623599     1